% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.fill}
\alias{aba.fill}
\title{\strong{aba.fill() - Fill in missing data to complete matrix of aba.query() results}}
\usage{
aba.fill(.data, db = gtx::config_db(), impala = getOption("gtx.impala",
  NULL))
}
\arguments{
\item{.data}{}

\item{db}{[Default = gtx::config_db()]}
}
\value{
data.frame
}
\description{
The \code{\link{aba.query}} will return only positive hits. Use 
this function to fill in the matrix of missing data based on positive hits.
}
\examples{
Basic use:
colocs_pos  <- aba.query(hgncid = "foo")
colocs_full <- aba.fill(colocs_pos)
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
