% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.filter}
\alias{aba.filter}
\title{\strong{aba.filter() - Recommended filtering for aba results}}
\usage{
aba.filter(.data, p12_ge = 0.8, minpval1_le = 1e-04,
  minpval2_le = 5e-08,
  log10_min_pval_colocGWAS_over_TopHitGWAS_lt = 0.5,
  gwas2gene_dist_lt = 1e+06, ncase_ge = 200, ncohort_ge = 200,
  protein_coding_only = TRUE, neale_only = FALSE, gsk_only = FALSE,
  fill_matrix = TRUE, flog_level = getOption("gtx.flog_level", "WARN"))
}
\arguments{
\item{.data}{\code{\link{aba.query_locus}} results object to filter}

\item{p12_ge}{[Default >= 0.80] This is the "H4" posterior probability cutoff}

\item{log10_min_pval_colocGWAS_over_TopHitGWAS_lt}{[Default < 0.5] Log10 ratio of coloc-gwas/top-hit-gwas pval. 
This ensures that the coloc gwas and gwas top hits are 1/2 order of magnitude apart and therefore shared.}

\item{gwas2gene_dist_lt}{[Default = 1e6] Max distance between coloc gene and gwas top hit.}

\item{neale_only}{[Default = FALSE] Filter onyl for Neale traits, reduces redundancy b/w GSK & Neale data.}

\item{gsk_only}{[Default = FALSE] Filter only for GSK traits, reduces redundancy b/w GSK & Neale data.}

\item{fill_matrix}{[Default = TRUE] Fill in the matrix based filtering. \emph{required} for \code{\link{aba.plot}}}

\item{flog_level}{[Default = getOption("gtx.flog_level", "WARN")] \code{\link{futile.logger}} \code{\link{flog.threshold}} [INFO, WARN, ERROR]}

\item{minpval1_lt}{[Default <= 1e-4] Min pval seen in the eQTL}

\item{minpval2_lt}{[Default <= 5e-8] Min pval seen in the GWAS data}

\item{ncase_lt}{[Default >= 200] Minimum ncases for traits}

\item{ncohort_lt}{[Default >= 200] Minimum ncohort for traits}

\item{protein_coding}{[Default = TRUE] Filter only for protein coding transcripts}
}
\value{
data.frame with input \code{\link{aba.query_locus}} reasonably filtered
}
\description{
The aba results and region queries will return colocs
that are not robust and therefore need further filtering.
This function provdies recommended minimal filtering to 
improve the robustness of the aba results. In addition, 
the results will fill in the matrix of positive colocs
based on the filtering to enable complete plotting. The
ability to fill in the matrix can be toggled.
}
\examples{
\\strong{Basic use:}
\\code{colocs_filtered <- aba.filter(colocs)}
\\code{colocs_filtered <- aba.filter(colocs, p12 = 0.9, minpval1 = 5e-8)}

\\strong{String together}
\\code{colocs_filtered <- aba.query_locus(hgncid = "HMGCR") \%>\% aba.filter()}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
