% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.flatten}
\alias{aba.flatten}
\title{\strong{aba.flatten() - Flatten full aba results}}
\usage{
aba.flatten(.data)
}
\arguments{
\item{.data}{\code{\link{aba.query}} results object to filter}
}
\value{
data.frame with input \code{\link{aba.query}} flattened to 1 gene / row.
}
\description{
The aba results and region queries will return colocs
}
\examples{
Basic use:
colocs  <- aba.query(hgncid = "foo")
colocs_flat <- aba.flatten(colocs)
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
