% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.info_flat}
\alias{aba.info_flat}
\title{\strong{aba.info_flat() - description of the cols from \code{\link{aba.query_locus}}}}
\usage{
aba.info_flat()
}
\value{
data.frame
}
\description{
\strong{aba.info_flat() - description of the cols from \code{\link{aba.query_locus}}}
}
\examples{
\\code{info <- aba.info_flat()}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
