% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.info_opt}
\alias{aba.info_opt}
\title{\strong{aba.info_opt() - info about initiating and setting global opts for aba}}
\usage{
aba.info_opt()
}
\value{
data.frame
}
\description{
\strong{aba.info_opt() - info about initiating and setting global opts for aba}
}
\examples{
aba.info_opt()
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
