% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.int_coloc_plot}
\alias{aba.int_coloc_plot}
\title{\strong{aba.int_coloc_plot() - Plot aba results}}
\usage{
aba.int_coloc_plot(.data, p12_ge = 0.8, max_dot_size = 5, title = NULL)
}
\arguments{
\item{p12_ge}{[Default = 0.80] Values below this will be colored as non-significant, while above have color for direction.}

\item{max_dot_size}{[Default = 5] Adjust max size of dots.}

\item{title}{Set the plot title. NULL will remove title.}
}
\value{
ggplot2 object for viz and export.
}
\description{
This function is for internal aba plotting use.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
