% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.plot}
\alias{aba.plot}
\title{\strong{aba.plot() - Plot aba results}}
\usage{
aba.plot(.data, ...)
}
\arguments{
\item{p12_ge}{[Default = 0.80] Values below this will be colored as non-significant, while above have color for direction.}

\item{max_dot_size}{[Default = 5] Adjust max size of dots.}

\item{title}{Set the plot title. NULL will remove title.}
}
\value{
ggplot2 object for viz and export.
}
\description{
This function tries to create a bubble plot to 
illustrate the coloc region. This function returns
a ggplot object which can be viewed or saved. It 
is \emph{highly} recommended to save the plot to pdf 
for best visualization of the plot.
}
\examples{
Basic use:
coloc_plot <- aba.plot(colocs)
coloc_plot <- aba.plot(colocs, max_dot_size = 3, title = "Cool plot")

View the returned object:
coloc_plot

Save the object to PDF:
ggsave(filename = "coloc_aba.pdf", 
       plot     = coloc_plot, 
       width    = 11,  
       height   = 8.5, 
       dpi      = 300)
       
Advanced use: query data, remove death related traits, filter, and then plot
colocs <- aba.wrapper(hgncid = "HMGCR")
colocs \%>\% filter(input == "HMGCR") \%>\% pluck("figures", 1) + ggtitle("Best plot ever")
colocs \%>\% filter(input == "HMGCR") \%>\% pluck("data", 1) \%>\% filter(hgncid != "bad_gene") \%>\% aba.plot()
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
