% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.query}
\alias{aba.query}
\title{aba.query() - Query aba results for all positive results}
\usage{
aba.query(analysis_ids, hgncid, ensemblid, rsid, chrom, pos, pos_start,
  pos_end, p12_ge = 0.8, minpval1_le = 1e-04, minpval2_le = 5e-08,
  surround = 1e+06, ncase_ge = 200, ncohort_ge = 200,
  protein_coding_only = FALSE, neale_only = FALSE, gsk_only = FALSE,
  db = gtx::config_db(), impala = getOption("gtx.impala", NULL))
}
\arguments{
\item{analysis_ids}{ukbiobank analysis id(s), single string or vector.}

\item{hgncid}{HGNC symbol. Single string or vector.}

\item{ensemblid}{Ensembl gene ID. Single string or vector.}

\item{rsid}{SNP rsid. Single string or vector.}

\item{chrom}{chromosome - Used to define a specific region}

\item{pos}{position that will have +/- \code{surround} for bounds to pull in colocs.}

\item{pos_start}{start position - Used to define a specific region, overrides surround}

\item{pos_end}{end position - Used to define a specific region, overrides surround}

\item{p12_ge}{[Default >= 0.80] This is the "H4" posterior probability cutoff}

\item{minpval1_le}{[Default <= 1e-4] Min pval seen in the eQTL}

\item{minpval2_le}{[Default <= 5e-8] Min pval seen in the GWAS data}

\item{surround}{[Default = 1e6] Distance to pull in TH + respective genes.}

\item{ncase_ge}{[Default >= 200] Minimum ncases for traits.}

\item{ncohort_ge}{[Default >= 200] Minimum ncohort for traits.}

\item{protein_coding_only}{[Default = TRUE] Filter only for protein coding transcripts}

\item{neale_only}{[Default = FALSE] Filter onyl for Neale traits, reduces redundancy b/w GSK & Neale data.}

\item{gsk_only}{[Default = FALSE] Filter only for GSK traits, reduces redundancy b/w GSK & Neale data.}

\item{db}{[Defalt = gtx::config_db()] Database to pull data from.}

\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}
}
\value{
data.frame with all coloc results in the region
}
\description{
Query all colocalization data in a region or set of traits. Typical use 
is to query a gene ID (\code{ensemblid} or \code{hgncid}) 
and to pull all coloc data in the surrounding queried 
region (+/- \code{surround} bp). A region can also be specified
using: \code{chrom}, \code{pos_start}, & \code{pos_end}. Alternatively,
a set of traits can be queried instead of the entire aba to help
generate hypothesis before looking at specific loci. The 
\code{\link{aba.query()}} will return only positive results based on
filtering (using params below).
}
\examples{
Query aba colocs:
colocs <- aba.query(hgncid = "HMGCR")

colocs <- aba.query(analysis_ids = "ukb_cool_analysis_id")
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
