% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.query_locus}
\alias{aba.query_locus}
\title{Coloc all-by-all functions}
\usage{
aba.query_locus(hgncid, ensemblid, surround = 1e+06, chrom, pos_start,
  pos_end, rsid, sc = getOption("gtx.sc", NULL),
  flog_level = getOption("gtx.flog_level", "WARN"))
}
\arguments{
\item{hgncid}{HGNC gene ID}

\item{ensemblid}{Ensembl gene ID}

\item{surround}{[Default = 1e6] Define region as \code{surround} +/- gene TSS.}

\item{chrom}{chromosome - Used to define a specific region}

\item{pos_start}{start position - Used to define a specific region, overrides surround}

\item{pos_end}{end position - Used to define a specific region, overrides surround}

\item{rsid}{SNP rsid}

\item{sc}{[Default =  getOption("gtx.sc")]. Spark connection.}

\item{flog_level}{[Default = getOption("gtx.flog_level", "WARN")] \code{\link{futile.logger}} \code{\link{flog.threshold}} [INFO, WARN, ERROR]}
}
\value{
data.frame with all coloc results in the region
}
\description{
\strong{aba.query_locus() - Query a locus for coloc all-by-all results}
Query all colocalization data in a region. Typical use 
is to query a gene ID (\code{ensemblid} or \code{hgncid}) 
and to pull all coloc data in the surrounding the queried 
gene (+/- \code{surround} bp). Alternatively, can specifiy 
the region coordinates using: \code{chrom}, \code{pos_start}, 
& \code{pos_end}. The rate limiting step is the \code{aba.query_locus}, 
after which \code{aba.filter} & \code{aba.plot} are very fast. 
This means \code{aba.query_locus} results should be saved to an object,
but the following steps can be piped together to quickly iterate 
filtering and ploting. \code{aba.query_locus()} takes ~ 7 min.
}
\examples{
\\strong{Query aba colocs:}
\\code{colocs <- aba.query_locus(hgncid = "HMGCR")}
\\code{colocs <- aba.query_locus(chrom = 1, pos_start = 1e7, pos_end = 2e7, sc = sc)}

\\strong{To establish a spark connection:}
\\code{sc <- spark_connect(master     = \\"yarn-client\\",
                    spark_home = \\"/opt/cloudera/parcels/SPARK2/lib/spark2\\",
                    version    = \\"2.1\\")}
\\strong{Set global gtx.sc option:}               
\\code{sc <- spark_connect(master     = \\"yarn-client\\",
                    spark_home = \\"/opt/cloudera/parcels/SPARK2/lib/spark2\\",
                    version    = \\"2.1\\")
     options(gtx.sc = sc)
     colocs <- aba.query_locus(hgncid = "HMGCR")}       
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
