% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.query_locus_wrapper}
\alias{aba.query_locus_wrapper}
\title{\strong{aba.query_locus_wrapper() - single function to wrap around: [\code{\link{aba.query_locus}}, \code{\link{aba.filter}}, \code{\link{aba.plot}}]}}
\usage{
aba.query_locus_wrapper(hgncid, ensemblid, surround = 1e+06, chrom, pos_start,
  pos_end, rsid, sc = getOption("gtx.sc", NULL),
  flog_level = getOption("gtx.flog_level", "WARN"), ...)
}
\arguments{
\item{hgncid}{HGNC gene ID}

\item{ensemblid}{Ensembl gene ID}

\item{surround}{[Default = 1e6] Define region as \code{surround} +/- input gene TSS.}

\item{chrom}{chromosome - Used to define a specific region}

\item{pos_start}{start position - Used to define a specific region, overrides surround}

\item{pos_end}{end position - Used to define a specific region, overrides surround}

\item{rsid}{SNP rsid}

\item{sc}{Spark connection}

\item{flog_level}{[Default = getOption("gtx.flog_level", "WARN")] \code{\link{futile.logger}} \code{\link{flog.threshold}} [INFO, WARN, ERROR]}

\item{...}{Pass filtering options to \code{\link{aba.filter}}}
}
\value{
ggplot2 object for viz and export.
}
\description{
This function is a simple wrapper around \code{\link{aba.query_locus}}, 
\code{\link{aba.filter}}, & \code{\link{aba.plot}}. This function 
is ideal for high throughput generation of many plots. This 
fucntion is \emph{not} ideal for deep dives into a locus.
}
\examples{
 
Basic use:
hmgcr_fig <- aba.wrapper(hgncid = "HMGCR")

Advanced use: TODO
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
