% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.query_traits}
\alias{aba.query_traits}
\title{\strong{aba.query_traits() - Query gwas traits for coloc all-by-all results}
Query a list of GWAS traits for eQTL coloc results.}
\usage{
aba.query_traits(analysis_ids, aba_filter = TRUE, flatten_data = TRUE,
  surround = 1e+06, protein_coding_only = TRUE,
  progressive_validation = FALSE, sc = getOption("gtx.sc", NULL),
  flog_level = getOption("gtx.flog_level", "WARN"), ...)
}
\arguments{
\item{analysis_ids}{ukbiobank analysis id(s), single string or vector.}

\item{aba_filter}{[Default = TRUE] Apply \code{\link{aba.filter}} to help validate colocs.}

\item{flatten_data}{[Default = TRUE] Condense coloc data into single gene per row.}

\item{surround}{[Default = 1e6] Distance from GWAS top hit to TSS.}

\item{protein_coding_only}{[Default = TRUE] Filtering to only return protein coding transcripts.}

\item{progressive_validation}{[Default = FALSE] TRUE = iteratively check data is obtained from the multiple queries. This dramatically slows down analysis, but is useful for trouble shooting.}

\item{sc}{[Default =  getOption("gtx.sc")]. Spark connection.}

\item{flog_level}{[Default = getOption("gtx.flog_level", "WARN")] \code{\link{futile.logger}} \code{\link{flog.threshold}} [INFO, WARN, ERROR]}

\item{...}{Options to pass to \code{\link{aba.filter}}}
}
\value{
a list with 2 data.frames. "flat" = flatten_data compact, 1 gene / row. "full" = full coloc results (after filtering)
}
\description{
\strong{aba.query_traits() - Query gwas traits for coloc all-by-all results}
Query a list of GWAS traits for eQTL coloc results.
}
\examples{
traits <- c("aid1", "aid2", "aid3")
traits_colocs <- aba.query_traits(analysis_ids = traits)
traits_colocs$flat \%>\% filter(hgncid == "Gene1")
traits_colocs$full \%>\% filter(hgncid == "Gene1")
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
