% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.save}
\alias{aba.save}
\title{aba.save - save \code{aba.wrapper} figures}
\usage{
aba.save(.data, path = getwd(), suffix = "aba-colocs", ...)
}
\arguments{
\item{.data}{Results from aba.wrapper or aba.plot}

\item{path}{[Default = \code{getwd}] Specify the path to save the figures}

\item{suffix}{[Default = "aba-colocs"] File name = .data$input_$suffix.pdf}

\item{...}{Parameters to pass to ggsave. e.g. dpi = 300}
}
\description{
This function will iterate over all inputs in an aba.wrapper results 
and save each figures using the input to name the output figure.
}
\examples{
Query aba colocs:
colocs <- aba.wrapper(hgncid = c("gene1", "gene2"))

The wraper returns 2 figures, aba.save() will save both figures
aba.save(colocs) 
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
