% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.wrapper}
\alias{aba.wrapper}
\title{aba.wrapper - single function to query and plot the aba colocs}
\usage{
aba.wrapper(analysis_ids, hgncid, ensemblid, rsid, chrom, pos, pos_start,
  pos_end, impala = getOption("gtx.impala", NULL), ...)
}
\arguments{
\item{analysis_ids}{ukbiobank analysis id(s), single string or vector.}

\item{hgncid}{HGNC symbol. Single string or vector.}

\item{ensemblid}{Ensembl gene ID. Single string or vector.}

\item{rsid}{SNP rsid. Single string or vector.}

\item{chrom}{chromosome - Used to define a specific region}

\item{pos_start}{start position - Used to define a specific region, overrides surround}

\item{pos_end}{end position - Used to define a specific region, overrides surround}

\item{surround}{[Default = 1e6] Distance from input gene/rsid to GWAS top hits.}

\item{p12_ge}{[Default >= 0.80] This is the "H4" posterior probability cutoff}

\item{minpval1_lt}{[Default <= 1e-4] Min pval seen in the eQTL}

\item{minpval2_lt}{[Default <= 5e-8] Min pval seen in the GWAS data}

\item{gwas2gene_dist_lt}{[Default = 1e6] Max distance between coloc gene and gwas top hit.}

\item{ncase_ge}{[Default >= 200] Minimum ncases for traits.}

\item{ncohort_ge}{[Default >= 200] Minimum ncohort for traits.}

\item{protein_coding}{[Default = TRUE] Filter only for protein coding transcripts}

\item{neale_only}{[Default = FALSE] Filter onyl for Neale traits, reduces redundancy b/w GSK & Neale data.}

\item{gsk_only}{[Default = FALSE] Filter only for GSK traits, reduces redundancy b/w GSK & Neale data.}
}
\value{
data.frame with the inputs used, all the data for each input, and default plots
}
\description{
aba.wrapper - single function to query and plot the aba colocs
}
\examples{
Query aba colocs:
colocs <- aba.wrapper(hgncid = "HMGCR")
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
