% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{add_ssh_known_host}
\alias{add_ssh_known_host}
\title{add_ssh_known_host}
\usage{
add_ssh_known_host(known_host)
}
\arguments{
\item{known_host}{A host to search for and add as known host. e.g. "github.com"}
}
\description{
This function adds a url to ssh known hosts. This allows us to enable users 
 to run some code (a module) to automatically do git updates without the user 
 needing to manually uthenticate that they want to add the url as a known host (e.g. github.com)
 when you do a git query (fetch/pull/etc).
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
