% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{annot}
\alias{annot}
\title{Query analysis, variant, gene etc. annotation}
\usage{
annot(analysis, chrom, pos, ref, alt, rs, hgnc, ensg, max_variants = 100L,
  max_genes = 10L, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{analysis}{Analysis identifier}

\item{chrom}{Chromosome identifier (string)}

\item{pos}{Position (on chromosome)}

\item{ref}{Reference allele}

\item{alt}{Alternate allele}

\item{rs}{dbSNP identifier (string e.g. rs123456)}

\item{hgnc}{HGNC gene symbol}

\item{ensg}{Ensembl gene identifier (string e.g. ENSG00000123456)}

\item{max_variants}{Maximum number of variants to return}

\item{max_genes}{Maximum number of genes to return}

\item{dbc}{Database connection (see \code{\link{gtxconnect}})}
}
\value{
Query results
}
\description{
Various standard queries of the tables that annotate the actual GWAS results
}
\author{
Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
