\name{clinical.import}
\alias{clinical.import}
\title{Import multiple files exported from SAS}
\description{
  Provides a convenience function to build a list of data frames,
  where each data frame is generated by reading a .txt file assumed
  to have been generated by SAS proc export.
}
\usage{
clinical.import(d, pattern = "^[a-zA-Z][a-zA-Z1-9]*\\\\.txt",
                usubjid = getOption("clinical.usubjid"),
                verbose = TRUE,
                only)
}
\arguments{
  \item{d}{Path to directory containing files exported from SAS}
  \item{pattern}{Regular expression for files to be imported}
  \item{usubjid}{Name of variable used for unique subject identifier}
  \item{verbose}{Whether to print progress messages}
  \item{only}{Character vector}
}
\details{
  Read the code or see the examples.
} 
\examples{
\dontrun{
clindata <- clinical.import("path/to/clinical/export/")
data(derivations.standard)
gxvars <- clinical.derive(clindata, derivations.standard)
summary(gxvars)
}
}
\author{
  Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
