% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_connections.R
\name{close_int_conn}
\alias{close_int_conn}
\title{Close connections they were internally created}
\usage{
close_int_conn(conn)
}
\arguments{
\item{conn}{Connection to check}
}
\description{
\strong{close_int_conn() - Close internally created connections}
This will close implyr or sparklyr connections that were internally created.
}
\examples{
close_int_conn(conn = impala)
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
