% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{config_db}
\alias{config_db}
\title{config_db}
\usage{
config_db(file = "~/.gtx_config.csv")
}
\arguments{
\item{file}{[Default = ~/.gtx_config.csv] Path to the config file to read in}
}
\value{
String of the database name
}
\description{
\strong{config_db - Read default project config database name}
Read in a default config file and identify the named database to use. 
This is useful in different projects where you want the code to point to different databses.
The config file is csv, 2 cols (key, value), and config_db uses key=database
}
\examples{
gtxconnection(use_database = config_db())
gtxconnection(use_database = config_db(file = "/some/path/config.csv"))
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
