% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{config_tmp_write_db}
\alias{config_tmp_write_db}
\title{config_tmp_write_db}
\usage{
config_tmp_write_db(file = "~/.gtx_config.csv")
}
\arguments{
\item{file}{[Default = ~/.gtx_config.csv] Path to the config file to read in}
}
\value{
String of the database name
}
\description{
\strong{config_tmp_write_db - Read default project config database name to write temporary tables to.}
Read in a default config file and identify the named database to use to write temporary tables to. 
This is useful in different projects where you want the code to point to different databses.
The config file is csv, 2 cols (key, value), and config_tmp_write_db uses key=tmp_write_db
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
