% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf.R
\name{credset}
\alias{credset}
\title{Determine whether a SNP is in the credible set}
\usage{
credset(bf, cred = 0.95)
}
\arguments{
\item{bf}{Vector of probabilites/frequencies}

\item{cred}{threshold for credible set (e.g. .95)}
}
\value{
ps	Vector of which values are in the credible set
}
\description{
Given a vector of probabilities bf, calculates which ones are in the credible set,
e.g. those for which the cumulative distribution is <.95
}
\author{
Unknown - Toby?
}
