\name{derivations.standard.IDSL}
\alias{derivations.standard.IDSL}
\alias{definitions.standard.IDSL}
\docType{data}
\title{Rules for deriving variables from IDSL clinical data}
\description{
  Standard rules for deriving variables for genetic association analysis
  from clinical data in IDSL format, for use with clinical.derive and associated functions.
}
\usage{data(derivations.standard.IDSL)}
\format{
  A data frame.

  You should not manipulate this data frame directly.  Instead, use \code{\link{derivation.add}()}.
}
\examples{
data(derivations.standard.IDSL)
}
\keyword{datasets}
