\name{derivations.standard}
\alias{derivations.standard}
\docType{data}
\title{Standard rules for deriving genetic association analysis
  variables from clinical data, v1}
\description{
A set of derivation rules for use with clinical.derive and associated functions
}
\usage{data(derivations.standard)}
\format{
  A data frame.

  You should not manipulate this data frame directly.  Instead, use \code{\link{derivation.add}()}.
}
\examples{
data(derivations.standard)
}
\keyword{datasets}
