% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_connections.R
\name{drop_impala_copy}
\alias{drop_impala_copy}
\title{Drop (tmp) impala_copy_to tables.}
\usage{
drop_impala_copy(.table = NULL)
}
\arguments{
\item{.table}{Table reference to drop}
}
\description{
\strong{drop_impala_copy() - drop tmp tables}
This function will easily drop \code{tbl} referenecs that were created by \code{impala_copy_to}.
This function will *not* drop other tables.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
