% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{extract_exposure}
\alias{extract_exposure}
\title{Extracts top instruments based on the chosen p-value (the instruments are distance-pruned)}
\usage{
extract_exposure(p = 5e-08, analyses = NULL, str = "",
  rsid = FALSE, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{p}{(optional): p-value threshold for the top findings where 5e-08 (default). Note! the instruments with a more lenient p-value threshold cannot be extracted.}

\item{analyses}{(optional): the study information: at least should have a column with analysis IDs; entity_type IDs if applicable.}

\item{str}{(optional): a keyword/keyphrase to choose the studies of interest; by default searches all studies}

\item{rsid}{(optional): logical; if FALSE (default) returns only chrom:position; if TRUE includes rsIDs (Warning! SNPs without rsIDs may not be returned);}

\item{dbc}{(default): other parameters}
}
\value{
the dataframe with exposure instruments (Warning! they have to be formatted and possibly clumped to be used in the MR analysis)
}
\description{
This function allows the users to pre-process and extract the 
required data from the GWA studies database, which then
can be used to run Mendelian Randomization and other related
analyses.
}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
