% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{extract_exposure_qtl}
\alias{extract_exposure_qtl}
\title{Extracts top instruments based on the chosen p-value (the instruments are not pruned)}
\usage{
extract_exposure_qtl(pval_thresh = 5e-08, analysis = NULL,
  entities = NULL, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{pval_thresh}{(optional): p-value threshold for the top findings where 5e-08 (default).}

\item{analysis}{(required): a string with analysis ID as defined in the GWA studies database}

\item{entities}{(optional): a dataframe which should have a column 'entity'; if NULL (default), all
entities for a specified analysis ID will be considered for instrument extraction}

\item{dbc}{(default): other parameters}
}
\value{
the dataframe with exposure instruments (Warning! they have to be formatted and clumped to be used in the MR analysis)
}
\description{
This function allows the users to extract the instruments
for molecular traits (e.g. transcripts or proteins) from the 
GWA studies database, which then can be used to run Mendelian 
Randomization and other related analyses.
}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
