% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{extract_outcome}
\alias{extract_outcome}
\title{Extracts the outcome instruments based on the exposure SNPs}
\usage{
extract_outcome(snp_list = NULL, analyses_exp = NULL, analyses_out,
  p = 5e-08, rsid = FALSE, search = "pos:chrom",
  dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{snp_list}{(required if the exposure instruments extracted from the custom file/existing dataframe): the dataframe which at least should have either two columns position (1) and chromosome (2)
or one column with rsIDs.}

\item{analyses_exp}{(required if the exposure instruments extracted from the database): the exposure study information; should have at least one column: analysis and if applicable: entity.}

\item{analyses_out}{(required): the outcome study information; should have at least one column: analysis and if applicable: entity.}

\item{p}{(optional): the p-value threshold to extract the exposure instruments from the database (if analyses_exp is not NULL).}

\item{rsid}{(optional): logical; if FALSE (default) returns SNPs identified only by chrom:position; if TRUE includes rsIDs (Warning! SNPs without rsIDs may not be returned);}

\item{search}{(optional): if "pos:chrom" (default) searches the instruments in the database by their chromosome position; alternatively, searches by their rsIDs (rsid should be "TRUE")
(only applicable to the searches for external files with the exposure instruments, i.e. snp_list is not NULL).}

\item{dbc}{(default): other parameters}
}
\value{
the dataframe with outcome instruments
}
\description{
This function allows the users to pre-process and extract the 
required data from the GWA studies database, which then
can be used to run Mendelian Randomization and other related
analyses.
}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
