% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{extract_study_info}
\alias{extract_study_info}
\title{Search study by keyword.}
\usage{
extract_study_info(str, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{str}{(required): a keyword/keyphrase to search for the study of interest}

\item{dbc}{(default): other parameters}
}
\value{
the study meta-information which then can be used in extract_outcome()
or extract_exposure() methods

internal
}
\description{
Finds a unique identifier for the study using a keyword or keyphrase. NOTE! keyword or keyphrase search is quite naive at this moment and may not return desirable results.
}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
