% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{filter_entities}
\alias{filter_entities}
\title{Choose the requested entities only}
\usage{
filter_entities(res = NULL, analyses)
}
\arguments{
\item{res}{(required): an output of the database query}

\item{analyses}{(required): a dataframe of requested studies}
}
\value{
a filtered query output with only entities indicated in 'analyses'
}
\description{
Removes the entities which were not in the list of requested studies.
}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
\keyword{internal}
