% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_qtl_bygene.R
\name{find_qtl_bygene}
\alias{find_qtl_bygene}
\title{Given an Ensembl ID, identify all significant QTLs}
\usage{
find_qtl_bygene(geneId, cutoff = 2e-05, dbc = getOption("gtx.dbConnection",
  NULL))
}
\arguments{
\item{geneId}{Gene to search. Recommended to use Ensembl ID. The function will try to infer gene symbols, but result may not be correct.}

\item{cutoff}{Only return QTLs below this p-value threshold}
}
\description{
This is an utility query to find all the QTLs for a specific gene.
It assumes that all the analyses for which entity_type is not null are QTLs
}
