% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flog_glued.R
\name{flog_debug}
\alias{flog_debug}
\title{flog_debug}
\usage{
flog_debug(.msg)
}
\arguments{
\item{.msg}{msg to glue and flog}
}
\description{
Simple wrapper around futile.logger::flog.debug(glue::glue(.msg)).
For full control over flog or glue opts use full explicit code.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
