% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flog_glued.R
\name{flog_error}
\alias{flog_error}
\title{flog_error}
\usage{
flog_error(.msg)
}
\description{
Simple wrapper around futile.logger::flog.error(glue::glue(.msg)).
For full control over flog or glue opts use full explicit code.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
