% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flog_glued.R
\name{flog_info}
\alias{flog_info}
\title{flog_info}
\usage{
flog_info(.msg)
}
\description{
Simple wrapper around futile.logger::flog.info(glue::glue(.msg)).
For full control over flog or glue opts use full explicit code.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
