% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flog_glued.R
\name{flog_warn}
\alias{flog_warn}
\title{flog_warn}
\usage{
flog_warn(.msg)
}
\description{
Simple wrapper around futile.logger::flog.warn(glue::glue(.msg)).
For full control over flog or glue opts use full explicit code.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
