% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{format_for_mr}
\alias{format_for_mr}
\title{Formats the data for MR analysis.}
\usage{
format_for_mr(data, type = "exposure", rsid = FALSE)
}
\arguments{
\item{data}{(required): the dataframe with the exposure or outcome instruments (output of: extract_exposure or extract_outcome)}

\item{type}{(optional): specifies the data type, i.e. 'exposure' (default) or outcome.}

\item{rsid}{(optional): logical; if FALSE (default) uses pos:chrom_ref_alt as SNP identifier; if TRUE uses rsIDs as SNP identifier.}
}
\value{
the formatted dataframe suitable for the MR analysis
}
\description{
Formats the database-extracted data for the Mendelian Randomization analysis (MR-Base)
}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
