% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{format_plot_labels}
\alias{format_plot_labels}
\title{Format plot labels for internal data}
\usage{
format_plot_labels(mrres, exp_flag = "internal", out_flag = "internal",
  dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{mrres}{(required): MR results as derived from run_mr_gsk()}

\item{exp_flag}{(optional): a flag that indicates a source of the exposure, i.e. internal or external;
labels are fomatted for the internal data only}

\item{out_flag}{(optional): a flag that indicates a source of the outcome, i.e. internal or external;
labels are formatted for the internal data only}
}
\value{
labels for the chosen exposure and outcome; if external, they will not be modified.
}
\description{
Format plot labels for internal data
}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
\keyword{internal}
