\name{gencode14.UGT1A1}
\alias{gencode14.UGT1A1}
\docType{data}
\title{GENCODE14 data for region around the UGT1A1 gene.}
\description{
  Subset of GENCODE14 for all transcripts partly or wholly in a region
  100kb around the UGT1A1 gene (specifically all transcripts on chr2
  with txEnd >= 234568893 and txStart <= 234781945).
}
\usage{
data(gencode14.UGT1A1)
}
\format{
  \code{gencode14.UGT1A1} is a data frame.
}
\source{
  \url{http://genome.ucsc.edu/cgi-bin/hgTables}
}
