% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{getUnprunedData}
\alias{getUnprunedData}
\title{Get Unpruned Data from the database}
\usage{
getUnprunedData(connectionType = "SQL", analysis, pval_thresh, maf_ge,
  rsq_ge, emac_ge, case_emac_ge, dbc)
}
\arguments{
\item{connectionType}{(character): The type of connection. Only SQL allowed for now}
}
\value{
A gwas S4 object
}
\description{
This functione is part of the gwas plotting functionality.
It retrives data that will ater be pruned
}
