% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtx_log.R
\name{gtx_fatal}
\alias{gtx_fatal}
\title{gtx_fatal}
\usage{
gtx_fatal(.msg)
}
\arguments{
\item{.msg}{msg to glue and flog}
}
\description{
Simple wrapper around futile.logger::flog.fatal(glue::glue(.msg)).
For full control over flog or glue opts use full explicit code.
}
\seealso{
Other gtx_log: \code{\link{gtx_debug}},
  \code{\link{gtx_error}}, \code{\link{gtx_info}},
  \code{\link{gtx_warn}}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
\concept{gtx_log}
