% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtx_log.R
\name{gtx_fatal_stop}
\alias{gtx_fatal_stop}
\title{gtx_fatal_stop}
\usage{
gtx_fatal_stop(.msg)
}
\arguments{
\item{.msg}{msg to glue and flog}
}
\description{
Wrapper around futile.logger::flog.fatal(glue::glue(.msg)), which
also throws a true fatal error with the same message, using 
stop().  Useful if calling within a function body where we want to 
stop execution, and also ensures the stop() message goes to
the console (even if futile.logger messages have been redirected).
}
\seealso{
Other gtx_log: \code{\link{gtx_debug}},
  \code{\link{gtx_error}}, \code{\link{gtx_fatal}},
  \code{\link{gtx_info}}, \code{\link{gtx_warn}}
}
\author{
Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
