% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{gtxcache}
\alias{gtxcache}
\title{Speed up search of genes and analyses}
\usage{
gtxcache(cache_analyses = TRUE, cache_genes = TRUE, disconnect = FALSE,
  dbc_from = getOption("gtx.dbConnection", NULL),
  dbc_to = getOption("gtx.dbConnection_cache", NULL))
}
\arguments{
\item{cache_genes}{Cache Gene coordinates}

\item{disconnect}{Disconnects the db. E.g. run gtxcache(disconnect=T) to free up memory}

\item{dbc_from}{DB to be cached (e.g. gwases)}

\item{dbc_to}{If there is already a gtxcache database available, use this instead of creating a new one}

\item{cache_analysis}{Cache the Analyses table}
}
\description{
Create a in-memory SQLite db storing analyses metadata and genes coordinates, to speed up searches
}
