% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtxbase.R
\name{gtxversion}
\alias{gtxversion}
\title{gtxversion - Report GTX package info}
\usage{
gtxversion()
}
\description{
Internal function to extract and report version and build information, 
including ref and sha1 if built from github using devtools.
Intent is for this to be printed to help time/versionstamp outputs
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
