% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas2.R
\name{gwas2}
\alias{gwas2}
\title{GWAS pairwise comparison}
\usage{
gwas2(analysis1, analysis2, pval1_thresh = 5e-08, pval2_thresh = 5e-08,
  prune_dist = 500000L, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{analysis1}{Analysis identifier for first GWAS (plot x-axis)}

\item{analysis2}{Analysis identifier for second GWAS (plot y-axis)}

\item{pval1_thresh}{Significance threshold for first GWAS. Default: 5e-08}

\item{pval2_thresh}{Significance threshold for second GWAS. Default: 5e-08}

\item{prune_dist}{Distance-based pruning threshold (for both).  Default: 500000}

\item{dbc}{Database connection. Default: getOption("gtx.dbConnection", NULL)}
}
\value{
Invisible dataframe of significant variants, with prune1/2 columns added
}
\description{
Compare results from two Genome Wide Association Studies.
}
\details{
Generates a plot showing (flipped) Z scores for all variants significant in either 
analysis.  Distance-based pruning is applied separately to each analysis, and 
index variants for first/second/both analyses are shown as red/green/yellow 
down-triangles/up-triangles/squares.  Using each set of index variants a GRS 
is constructed and tested for association in the other analysis.
}
\author{
Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
