\name{hapmap.snpdata}
\alias{hapmap.snpdata}
\title{Extract individual level snp/haplotype data from HapMap}
\description{
  Does this.
}
\usage{
hapmap.snpdata(params, hapmap)
}
\arguments{
  \item{params}{a parameterised risk score}
  \item{hapmap}{a hapmap list object}
}
\details{
See hapmap for how to build.  Returns a snpdata object with special ploidy=1.
}
\author{
  Toby Johnson \email{Toby.x.Johnson@gsk.com}
}