% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_get_states}
\alias{ideas_get_states}
\title{ideas_get_states}
\usage{
ideas_get_states(chrom, pos, staten, states = NULL, states_data,
  impala = getOption("gtx.impala", NULL), db = config_db())
}
\arguments{
\item{chrom}{Chromosome to query}

\item{pos}{Position}

\item{staten}{Number of states.}

\item{states}{[NULL]}

\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}

\item{db}{[config_db()] Database to use for queries.}
}
\description{
Find the closest IDEAS state(s) to a chrom:pos.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
