% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_gwas_plot}
\alias{ideas_gwas_plot}
\title{ideas_gwas_plot}
\usage{
ideas_gwas_plot(.data, path, cutoff_gt = 0.001, legend = TRUE)
}
\arguments{
\item{.data}{\code{\link{ideas_gwas_summarize}} output}

\item{path}{Path for output PDF.}

\item{cutoff_gt}{[0.001] Plotting "Minimum proportion of genetic signal" predicted in cell types. Range = o:1.}

\item{legend}{[TRUE] TRUE = display cell type legend}
}
\description{
Create a plot of the \code{\link{ideas_gwas_summarize}} data.
}
\seealso{
Other ideas_predict: \code{\link{ideas_gwas_summarize}},
  \code{\link{ideas_loci_summarize}},
  \code{\link{ideas_make}}, \code{\link{ideas_predict}},
  \code{\link{ideas_preload_states}},
  \code{\link{ideas_wrapper}}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
