% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_lasso}
\alias{ideas_lasso}
\title{ideas_lasso}
\usage{
ideas_lasso(chrom, pos, vindex, vprior, vlp, states = NULL,
  usegenomebackground = TRUE, inter = 5, alpha = 0.99, layer = 10,
  permute = FALSE, impala = getOption("gtx.impala", NULL))
}
\arguments{
\item{usegenomebackground}{[Default = TRUE]}

\item{inter}{[Default = 5]}

\item{alpha}{[Default = 0.99]}

\item{layer}{[Default = 10]}

\item{permute}{[Default = FALSE]}

\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}
}
\value{
TBD
}
\description{
Using gglasso, identify the most likely cell type per locus.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
