% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_loci}
\alias{ideas_loci}
\title{ideas_loci}
\usage{
ideas_loci(ideas_make_obj, ideas_predict_obj,
  impala = getOption("gtx.impala", NULL), db = config_db())
}
\arguments{
\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}

\item{db}{[Default = config_db()] Database to use for queries.}
}
\value{
TBD
}
\description{
IDEAS predict data for a locus.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
