% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_loci_summarize}
\alias{ideas_loci_summarize}
\title{ideas_loci_summarize}
\usage{
ideas_loci_summarize(ideas_make_dat, ideas_predict_dat)
}
\arguments{
\item{ideas_make_dat}{\code{\link{ideas_make}} object}

\item{ideas_predict_dat}{\code{\link{ideas_predict}} object}

\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}

\item{db}{[config_db()] Database to use for queries.}
}
\description{
IDEAS predict data for a locus.
}
\seealso{
Other ideas_predict: \code{\link{ideas_gwas_plot}},
  \code{\link{ideas_gwas_summarize}},
  \code{\link{ideas_make}}, \code{\link{ideas_predict}},
  \code{\link{ideas_preload_states}},
  \code{\link{ideas_wrapper}}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
\concept{ideas_predict}
