% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_plot}
\alias{ideas_plot}
\title{ideas_plot}
\usage{
ideas_plot(.data, path, cutoff = 0.001, group = FALSE, relaxed = NA,
  legend = TRUE, impala = getOption("gtx.impala", NULL),
  db = config_db())
}
\arguments{
\item{.data}{\code{ideas_predict} output}

\item{path}{Path for output PDF.}

\item{cutoff}{[Default = 0.001]}

\item{group}{[Default = FALSE]}

\item{relaxed}{[Default = NA]}

\item{legend}{[Default = TRUE]}

\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}

\item{db}{[Default = config_db()] Database to use for queries.}
}
\value{
TBD
}
\description{
Plot \code{ideas_predict}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
