% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_preload_states}
\alias{ideas_preload_states}
\title{ideas_preload_states}
\usage{
ideas_preload_states(impala = getOption("gtx.impala", NULL),
  db = config_db(), chrom = c(1:22, "X", "Y"))
}
\arguments{
\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}

\item{db}{[config_db()] Database to use for queries.}

\item{chrom}{[c(1:22, "X", "Y")] Specify which chromosomes to load.}
}
\value{
states_data to be used with \code{ideas_predict}
}
\description{
Preload all IDEA states across "all" chrom. This is ideal for when running ideas_predict across multiple GWAS;
however, it would be slower for single GWAS.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
