% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_sum_pp_cross_cell_types}
\alias{ideas_sum_pp_cross_cell_types}
\title{ideas_sum_pp_cross_cell_types}
\usage{
ideas_sum_pp_cross_cell_types(prior, prob, cellweight, snpindex,
  confidence = 0.95, cut = 0, layer = 10)
}
\arguments{
\item{prior}{TBD}

\item{prob}{TBD}

\item{cellweight}{TBD}

\item{snpindex}{TBD}

\item{confidence}{[0.95]}

\item{cut}{[0]}

\item{layer}{[10]}
}
\description{
Sum posterior probabilities across cell types
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
