% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_wrapper}
\alias{ideas_wrapper}
\title{ideas_preload_states}
\usage{
ideas_wrapper(analysis, ht_load = FALSE, states_data, relaxed = 3,
  group = FALSE, impala = getOption("gtx.impala", NULL),
  dbc = getOption("gtx.dbConnection", NULL), db = config_db())
}
\arguments{
\item{analysis}{A string or vector of analysis ids.}

\item{ht_load}{[FALSE] TRUE = Load the IDEAS states for high throughput.}

\item{states_data}{Pass \code{\link{ideas_preload_states}} data instead of loading it.}

\item{relaxed}{[3] Cell types count toward total if they are in the top # of 'relaxed' cell types. e.g., 3 = The top 3 cell types count.}

\item{group}{[FALSE] TRUE = group cell types together.}

\item{impala}{[getOption("gtx.impala", NULL)]}

\item{dbc}{[getOption("gtx.dbConnection", NULL)]}

\item{db}{[`gtx::config_db()`]}
}
\description{
Preload all IDEA states across "all" chrom. This is ideal for when running ideas_predict across multiple GWAS;
however, it would be slower for single GWAS.
}
\seealso{
Other ideas_predict: \code{\link{ideas_gwas_plot}},
  \code{\link{ideas_gwas_summarize}},
  \code{\link{ideas_loci_summarize}},
  \code{\link{ideas_make}}, \code{\link{ideas_predict}},
  \code{\link{ideas_preload_states}}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
\concept{ideas_predict}
