% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_connections.R
\name{impala_copy_to}
\alias{impala_copy_to}
\title{Copy data to user's home database for joins with impala tables.}
\usage{
impala_copy_to(df, dest = getOption("gtx.impala", NULL),
  database = gtx::config_tmp_write_db(), table_name = NULL,
  random_name = TRUE)
}
\arguments{
\item{df}{Data to copy to RDIP}

\item{dest}{Impala implyr connection}

\item{database}{Defaults to gtx::config_tmp_write_db().}

\item{table_name}{[Default = NULL] Specify table name, overrides random_name = TRUE. If random = FALSE & table_name is NULL, then defaults to tmp_whoami()_data4join}

\item{random_name}{[Default = TRUE] TRUE = tmp_whoami()_timestamp. FALSE & is.null(table_name) = tmp_whoami()_data4join.}
}
\value{
table reference to the \code{df} in RDIP
}
\description{
\strong{impala_copy_to() - copy data to RDIP}
This will copy a dataframe to the home directory of the user 
and return the
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
