% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{int_coloc_pics_lite}
\alias{int_coloc_pics_lite}
\title{int_coloc_pics_lite 
  Test for colocalization of two PICS sets}
\usage{
int_coloc_pics_lite(data1, data2, pics1 = "PICS_probability",
  pics2 = "PICS_probability", rsid1 = "Linked_SNP",
  rsid2 = "Linked_SNP", rounded = 6, priorc1 = 1e-04,
  priorc2 = 1e-04, priorc12 = 1e-05)
}
\arguments{
\item{data1, data2}{PICS sets from read.pics or download.pics}

\item{pics1, pics2}{column name to pull PICS prob from. Default = "PICS_probability"}

\item{rsid1, rsid2}{column name to pull rsid from.      Default = "Linked_SNP"}

\item{rounded}{Decimal points to round posteriors to}

\item{priorc1}{Prior probability for colocalization with siganl for data1  Default = 1e-4}

\item{priorc2}{Prior probability for colocalization with siganl for data2 Default = 1e-4}

\item{priorc12}{Prior probability for colocalization of both signals.   Default = 1e-5}
}
\value{
only H3 & H4 posteriors
}
\description{
int_coloc_pics_lite 
  Test for colocalization of two PICS sets
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
