% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{int_fastDoCall}
\alias{int_fastDoCall}
\title{An Alternative To The Internal Do.Call
 
  The do.call can be somewhat slow, especially when working with large objects. 
  This function is based upon the suggestions from Hadley Wickham on the R mailing list, 
  see here. Also thanks to Tommy at StackOverflow for suggesting how to handle double 
. and triple colon operators, ::, further enhancing the function.}
\usage{
int_fastDoCall(what, args, quote = FALSE, envir = parent.frame())
}
\arguments{
\item{what}{either a function or a non-empty character string naming the function to be called.}

\item{args}{a list of arguments to the function call. The names attribute of args gives the argument names.}

\item{quote}{a logical value indicating whether to quote the arguments.}

\item{envir}{an environment within which to evaluate the call. This will be most useful if what is a character string and the arguments are symbols or quoted expressions.}
}
\description{
An Alternative To The Internal Do.Call
 
  The do.call can be somewhat slow, especially when working with large objects. 
  This function is based upon the suggestions from Hadley Wickham on the R mailing list, 
  see here. Also thanks to Tommy at StackOverflow for suggesting how to handle double 
. and triple colon operators, ::, further enhancing the function.
}
\author{
Max Gorden
}
