% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{int_harmonize_pics}
\alias{int_harmonize_pics}
\title{int_harmonize_pics}
\usage{
int_harmonize_pics(data1, data2, opts = data.frame(pics1 =
  "PICS_probability", pics2 = "PICS_probability", rsid1 = "Linked_SNP",
  rsid2 = "Linked_SNP", stringsAsFactors = FALSE))
}
\description{
int_harmonize_pics
}
