% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{int_ht_phewas}
\alias{int_ht_phewas}
\title{int_ht_phewas}
\usage{
int_ht_phewas(hgnc, hgncid, rs, rsid, chrom, pos, ref, alt,
  ignore_ukb_neale = TRUE, ignore_ukb_cane = TRUE,
  ignore_qtls = TRUE, phewas_pval_le = 1e-07,
  dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{hgnc}{Will analyze all moderate/high impact VEP variants overlapping gene position (db.vep)}

\item{rs}{RSID(s) to analyze as a string or vector}

\item{chrom}{chromosome(s) to analyze as a string or vector}

\item{pos}{positions(s) to analyze as a single or vector}

\item{ref}{ref allele(s) to analyze as a string or vector. Optional w.r.t. chrom/pos.}

\item{alt}{alt allele(s) to analyze as a string or vector. Optional w.r.t. chrom/pos.}

\item{ignore_ukb_neale}{[Default = TRUE] TRUE = ignore Neale UKB GWAS in PheWAS}

\item{ignore_ukb_cane}{[Default = TRUE] TRUE = ignore Canelaxandri UKB GWAS in PheWAS}

\item{ignore_qtls}{[Default = TRUE] TRUE = ignore QTLs in PheWAS (gwas_results with entity)}

\item{dbc}{[Default = getOption("gtx.dbConnection", NULL)] gtxconnection.}
}
\value{
tibble with all PheWAS hits!
}
\description{
High-throughput PheWAS
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
