% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{int_ht_regional_annot}
\alias{int_ht_regional_annot}
\title{int_ht_regional_annot}
\usage{
int_ht_regional_annot(cs_df, chrom, pos, ref, alt)
}
\arguments{
\item{chrom}{chromosome (Mandatory)}

\item{pos}{position (Mandatory)}

\item{input}{data.frame with chrom, pos, analysis, signal; likely & suggested input from int_ht_phewas. Other cols will be kept.}

\item{analysis}{analysis ID (Mandatory)}

\item{signal}{CLEO signal OR NA. (Mandatory)}

\item{...}{Params to pass to regionplot.data (e.g. surround = 1e6)}

\item{dbc}{[Default = getOption("gtx.dbConnection", NULL)] gtxconnection.}
}
\value{
tibble
}
\description{
High-Throughput regional_contex annot of cs.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
