% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{int_ht_regional_context}
\alias{int_ht_regional_context}
\title{int_ht_regional_context}
\usage{
int_ht_regional_context(input, chrom, pos, ref, alt, analysis, signal,
  cpu = 8, drop_cs = TRUE)
}
\arguments{
\item{input}{data.frame with chrom, pos, ref, alt, analysis, signal; likely from phewas.}

\item{chrom}{chromosome}

\item{pos}{position}

\item{ref}{reference allele}

\item{alt}{alt allele}

\item{analysis}{analysis ID}

\item{signal}{CLEO signal}
}
\value{
tibble
}
\description{
High-Throughput regional_context.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
