% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{int_ht_regional_context_analysis}
\alias{int_ht_regional_context_analysis}
\title{int_ht_regional_context_analysis}
\usage{
int_ht_regional_context_analysis(input, style, cpu = 8, drop_cs = TRUE,
  ...)
}
\arguments{
\item{input}{int_ht_phewas object}

\item{style}{CLEO style {'signal'|'signals'}}

\item{cpu}{[Default=8]}
}
\value{
tibble
}
\description{
High-throughput regional_context_analysis
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
