% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{int_ht_rpd_wrapper}
\alias{int_ht_rpd_wrapper}
\title{int_ht_rpd_wrapper}
\usage{
int_ht_rpd_wrapper(analysis, chrom, pos, signal, use_database, ...)
}
\arguments{
\item{analysis}{analysis ID}

\item{chrom}{chromosome}

\item{pos}{position}

\item{signal}{CLEO signal or NA}

\item{use_database}{Database for gtxconnect}

\item{...}{Additional arguements for regionplot.data}
}
\value{
tibble with the credible set AND the variant queried
}
\description{
int_ht_rpd_wrapper - internal high-throughput RegionPlot.Data wrapper.
}
\examples{
ret <- tibble(analysis = "GSK500KV3lin_HES_Asthma", chrom = "9", pos = 6255967, signal = 4) \%>\%  
       mutate(cs = future_pmap(list(analysis, chrom, pos, signal), 
                               ~int_ht_rpd_wrapper(analysis = ..1, chrom = ..2, pos = ..3, 
                                                        signal = ..4, use_database = config_db())))
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
