% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{int_input_2_select_snps_sql}
\alias{int_input_2_select_snps_sql}
\title{int_input_2_select_snps_sql}
\usage{
int_input_2_select_snps_sql(hgnc, hgncid, rs, rsid, chrom, pos, ref, alt)
}
\arguments{
\item{hgnc}{Will analyze all moderate/high impact VEP variants overlapping gene position (db.vep)}

\item{rs}{RSID(s) to analyze as a string or vector}

\item{chrom}{chromosome(s) to analyze as a string or vector}

\item{pos}{positions(s) to analyze as a single or vector}

\item{ref}{ref allele(s) to analyze as a string or vector. Optional w.r.t. chrom/pos.}

\item{alt}{alt allele(s) to analyze as a string or vector. Optional w.r.t. chrom/pos.}
}
\value{
string to select variants
}
\description{
Get an impala table reference for the various inputs harmonized to chrom/pos/ref/alt
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
