% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{int_sbind}
\alias{int_sbind}
\title{Combine Shape Objects}
\usage{
int_sbind(x, y, fill = NA)
}
\description{
Combine shape objects into one shape object. It works analogous to rbind.
}
\author{
Martijn Tennekes
}
