% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{multicoloc.data_kbs}
\alias{multicoloc.data_kbs}
\title{multicoloc.data_kbs gets data like multicoloc.data but also cals colocs}
\usage{
multicoloc.data_kbs(analysis1, analysis2, chrom, pos_start, pos_end, pos,
  hgncid, ensemblid, rs, surround = 0, hard_clip = FALSE,
  dbc = getOption("gtx.dbConnection", NULL), ...)
}
\description{
multicoloc.data_kbs gets data like multicoloc.data but also cals colocs
}
