% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{multicoloc.plot_kbs}
\alias{multicoloc.plot_kbs}
\title{multicoloc.plot_kbs ggplot implementation of multicoloc plots.}
\usage{
multicoloc.plot_kbs(.data, thresh_analysis_ge = 0.5,
  thresh_entity_ge = 0.5)
}
\arguments{
\item{.data}{multicoloc.data_kbs() dataframe}

\item{thresh_analysis_ge}{[Default = 0.5]}

\item{thresh_entity_ge}{[Default = 0.5]}
}
\description{
multicoloc.plot_kbs ggplot implementation of multicoloc plots.
}
