% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{multicoloc2}
\alias{multicoloc2}
\title{Multiple colocalization, use case 2}
\usage{
multicoloc2(analysis1, target, chrom, pos_start, pos_end, pos, hgncid,
  ensemblid, rs, surround = 0, dbc = getOption("gtx.dbConnection",
  NULL), ...)
}
\arguments{
\item{analysis1}{analysis id to colocalize}

\item{target}{data frame with columns analysis and entity}

\item{chrom}{other arguments used to determine region for coloc}
}
\value{
A data frame of analysis1 against each analysis/entity from
 target (called analysis2 and entity2) along with 
 colocalization probabilities.
}
\description{
Multiple colocalization, one (disease) analysis against 
a set of analysis/entity pairs.
}
\details{
A second use case for multicoloc, is that a single analysis1 
(and entity1 in future) is to be colocalized with a potentially 
large set of analysis2 (with or without entity2; not necessarily 
e/pQTL).
The Motivating example is taking a top hit variant at a (disease) 
GWAS locus,PheWAS'ing it, and then testing colocalization with all 
the PheWAS hits.
}
\author{
Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
