% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_gene.R
\name{nearest_gene}
\alias{nearest_gene}
\title{nearest_gene}
\usage{
nearest_gene(analysis, hgnc, hgncid, ensemblid, rs, rsid, chrom, pos,
  nearest_method = c("nearest_tss", "nearest_dist_pruned",
  "nearest_localized"), dist_prune_ge = 50000,
  protein_coding_only = TRUE, include_negative_results = FALSE,
  surround = 1e+06, gwas_pval_le = 5e-08, calc_th = FALSE, gwas_args,
  aba_gwas_results = FALSE, ignore_ukb_neale = TRUE,
  ignore_ukb_cane = TRUE, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{analysis}{GWAS analysis ID}

\item{hgnc}{Analyze all moderate/high impact VEP variants overlapping gene position.}

\item{ensemblid}{Ensembl gene ID.}

\item{rs}{RSID(s) to analyze as a string or vector #Not implemented yet}

\item{chrom}{chromosome(s) to analyze as a string or vector #Not implemented yet}

\item{pos}{positions(s) to analyze as a single or vector  #Not implemented yet}

\item{nearest_method}{[Default = 'nearest_tss'] String(s) specifying the method for calculating nearest gene. 
{'nearest_tss', 'nearest_dist_pruned', 'nearest_localized'}}

\item{dist_prune_ge}{[Default = 50000] For nearest_method 'nearest_localized', this is the distance that the second nearest gene must be from the TH.}

\item{include_negative_results}{[Default = FALSE] Return non-nearest gene within surround distance.}

\item{surround}{[Default = 1e6] Max distance (bp) to non-nearest gene.}

\item{ignore_ukb_neale}{[Default = TRUE] TRUE = ignore Neale UKB GWAS in PheWAS}

\item{ignore_ukb_cane}{[Default = TRUE] TRUE = ignore Canelaxandri UKB GWAS in PheWAS}

\item{dbc}{[Default = getOption("gtx.dbConnection", NULL)] gtxconnection.}

\item{ignore_qtls}{[Default = TRUE] TRUE = ignore QTLs in PheWAS (gwas_results with entity)}
}
\value{
tibble (data.frame)
}
\description{
Calculate nearest_gene V2G to identify targets~indications.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
