% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{phewas}
\alias{phewas}
\alias{phewas.data}
\alias{phewas.qq}
\title{Phenome Wide Assocation Study}
\usage{
phewas(chrom, pos, ref, alt, rs, analysis, analysis_not,
  description_contains, phenotype_contains, has_tag, ncase_ge, ncohort_ge,
  analysis_fields = c("description", "label", "unit", "ncase",
  "ncontrol", "ncohort"), tag_is = "phewas_group", plot_ymax = 30,
  dbc = getOption("gtx.dbConnection", NULL))

phewas.data(chrom, pos, ref, alt, rs, analysis, analysis_not,
  description_contains, phenotype_contains, has_tag, ncase_ge, ncohort_ge,
  analysis_fields = c("description", "label", "unit", "ncase",
  "ncontrol", "ncohort"), tag_is, with_tags = FALSE, nearby,
  dbc = getOption("gtx.dbConnection", NULL))

phewas.qq(chrom, pos, ref, alt, rs, analysis, analysis_not,
  description_contains, phenotype_contains, ncase_ge, ncohort_ge,
  dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{chrom}{Chromosome of variant of interest}

\item{pos}{Position of variant of interest}

\item{ref}{Reference allele of variant of interest}

\item{alt}{Alternate allele of variant of interest}

\item{rs}{dbSNP rs identifier of variant of interest}

\item{analysis}{Identifiers of analyses to include}

\item{analysis_not}{Identifiers of analyses to exclude}

\item{description_contains}{Search term for analyses to include}

\item{has_tag}{Tag value for analyses to include}

\item{ncase_ge}{Threshold case sample size greater-or-equal for inclusion}

\item{ncohort_ge}{Threshold cohort sample size greater-or-equal for inclusion}

\item{analysis_fields}{Fields from analysis metadata to return}

\item{tag_is}{Tag flag to use for plot grouping}

\item{plot_ymax}{Y-axis value maximum}

\item{dbc}{Database connection}

\item{nearby}{Additionally return minimum pvalue at position+-nearby}
}
\value{
Data frame of summary statistics
}
\description{
Look up and plot results for a Phenome Wide Association Study.
}
\details{
Look up association summary statistics for a single variant of interest, specified 
either by chromosome and position, or by dbSNP rs identifier.  (FIXME need to add optional 
ref/alt arguments to disambiguate when there are multiple variants at the same position,
or multiple variants with the same rs.)

By default, summary statistics are returned for all analyses where the variant of interest 
was tested.  If results are required for only a specific set of analyses, these can be 
specified using a combination of arguments that are passed through to \code{gtxanalyses()}.
}
\section{Functions}{
\itemize{
\item \code{phewas.data}: Look up results without plotting

\item \code{phewas.qq}: Draw simple QQ plot of PheWAS results
}}

\author{
Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
