% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{piccolo}
\alias{piccolo}
\title{PICCOLO: Colocalization of GWAS and eQTL/pQTL signals without summary stats}
\usage{
piccolo(rs, pval, ancestry = NULL, indication = NULL,
  dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{rs}{SNP rsid or chr:pos id. Single string or vector.}

\item{pval}{the association p-value for the SNP. Single string or vector.}

\item{ancestry}{Ancestry information: EUR/ASN/AFR. Currently, only EUR is available. OPTIONAL!!}

\item{indication}{Associated phenotype or trait info. OPTIONAL!!}

\item{dbc}{Database connection. Default: getOption("gtx.dbConnection", NULL)}
}
\value{
a data frame containing the result of the
 PICS calculation.
}
\description{
PICCOLO estimates the probability that two genetic signals between GWAS and e/pQTLs 
. are shared using colocalization approach. A proper colocalization analysis requires 
. full summary stats for both the GWAS and QTL studies. Unfortunately, studies frequently 
  do not share full summary stats which makes a traditional colocalization analysis impossible..
. In this case, PICCOLO enable to do a colocalization using PICS.
}
\author{
Kijoung Song \email{kys21207@gsk.com}
}
