% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\name{pics_calc}
\alias{pics_calc}
\title{PICS calculation using the linkage information of 10,000 UKB samples (EUR only)}
\usage{
pics_calc(index.data, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{dbc}{Database connection. Default: getOption("gtx.dbConnection", NULL)}

\item{data}{Data frame with columns snpID, pval, ancestry and indication. Use the same column names for your data. However, ancestry and indication are optional.}
}
\value{
a data frame containing the result of the
 PICS calculation.
}
\description{
The PICS algorithm calculates the most likely causal SNPs given the observed association signal
. at a locus. For an associated locus, enter the most highly-associated SNP 
  (referred to as the index SNP) and the strength of association. Using the linkage information
  in 10,000 UKB samples, the algorithm identifies the SNPs that are most likely to be the causal variants
  responsible for the association (PICS_Probability).
}
\details{
See \strong{Genetic and Epigenetic Fine-Mapping of Causal Variants in Autoimmune Disease} 
  by Kyle Kai-How Farh,et al. Nature 518, 337–343 (19 February 2015) at
  \url{http://www.nature.com/nature/journal/v518/n7539/full/nature13835.html#close}
}
\author{
Kijoung Song \email{kys21207@gsk.com}
}
