% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\docType{methods}
\name{plot,gwas,missing-method}
\alias{plot,gwas,missing-method}
\title{Plotting function for gwas}
\usage{
\S4method{plot}{gwas,missing}(x, y, style, qqplot_alpha = NULL,
  qqplot_col = NULL, plot_fastbox = NULL)
}
\description{
Plotting function for gwas
}
