% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{prepareGwasPlot}
\alias{prepareGwasPlot}
\title{Prepare data from plotting manhattan and/or qqplot}
\usage{
prepareGwasPlot(gwasObj, dbc, analysis, maf_ge, rsq_ge, emac_ge,
  case_emac_ge, plot_fastbox, style)
}
\arguments{
\item{gwasObj}{(gwas): a gwas S4 object}
}
\value{
A gwas S4 class object with information that will be used by the plotting function
}
\description{
Prepare data from plotting manhattan and/or qqplot
}
