% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{pruneData}
\alias{pruneData}
\title{Function to prune data}
\usage{
pruneData(gwasObj, prune_dist, gene_annotate)
}
\arguments{
\item{gwasObj}{(gwas): a gwas S4 object}

\item{gene_annotate}{(logical): whether to annotate top hits with gene names}
}
\value{
The input gwas object with added pruned data
}
\description{
Function to prune data
}
