\name{replaceNA}
\alias{replaceNA}
\title{Replace NA with something else}
\description{
  Convenience function to replace NA with something else.
}
\usage{
replaceNA(x, replacement = NA)
}
\arguments{
  \item{x}{an object}
  \item{replacement}{what NA is to be replaced with}
}
\value{
  \code{x} with NA values replaced by \code{replacement}.
}
\author{
  Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
