% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{run_mr_gsk}
\alias{run_mr_gsk}
\title{Runs the Mendelian Randomization (MR) and sensitivity analyses}
\usage{
run_mr_gsk(exp, out, rsid = FALSE, clump = TRUE, hetero = TRUE,
  pleio = TRUE, single = TRUE)
}
\arguments{
\item{exp}{(required): the formatted exposure instruments (output of: format_for_mr())}

\item{out}{(required): the formatted outcome instruments (output of: format_for_mr())}

\item{rsid}{(optional): logical; if FALSE (default) uses pos:chrom_ref_alt as a SNP identifier; 
if TRUE uses rsIDs as a SNP identifier (the automated clumping is enabled for this option).}

\item{clump}{(optional): logical; if TRUE (default) enables an automated pruning of the exposure instruments 
(only applicable if rsid=TRUE); if FALSE disables an automated pruning of the exposure instruments.}

\item{hetero}{(optional): logical; enables/diables the heterogeneity testing (TRUE by default).}

\item{pleio}{(optional): logical; enables/disables the pleiotropy testing (TRUE by default).}

\item{single}{(optional): logical; enables/disables the single SNP MR analysis (TRUE by default).}
}
\value{
the list of dataframes: 
(1) MR results
(2) Heterogeneity analysis
(3) Pleiotropy analysis
(4) Single SNP MR analysis
(5) Harmonisation data
}
\description{
This function runs MR and sensitivity analyses (TwoSampleMR) using the pre-formatted datasets (either format_data (TwoSampleMR) or format_for_mr (gtx)) where SNPs should be uniquely identified by either chromosome positions & alleles or rsIDs.
}
\examples{
To access the dataframe, e.g. \dontrun{df[[1]]} - MR results

}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
