\name{snps.BRCA1}
\alias{snps.BRCA1}
\docType{data}
\title{Genotyping array annotation for SNPs near the BRCA1 gene.}
\description{
  Subset of the array annotation for the Illumina HumanOmniExpressExome
  array, at positions chr17:41,145,568-41,328,243, which covers (by a
  generous margin) all transcripts for the BRCA1 gene.
}
\usage{
data(snps.BRCA1)
}
\format{
  \code{snps.BRCA1} is a data frame.
}
\source{
  \url{http://www.illumina.com/support/array/array_kits/infinium_humanomniexpressexome_beadchip_kit/downloads.ilmn}
}
