% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionplot.R
\name{styleLd}
\alias{styleLd}
\title{S4 class containing the data to draw a regional association plot with the 
  'ld' style.}
\usage{
styleLd(pValues, chrom, pos_start, pos_end, pos, rs,
  dbc = getOption("gtx.dbConnection", NULL))
}
\description{
S4 class containing the data to draw a regional association plot with the 
  'ld' style.
}
\section{Slots}{

\describe{
\item{\code{ldPValues}}{Data frame containing the p-values for the 'ld' style.}

\item{\code{indexVariant}}{Data frame with one row, containing the index variant.}
}}

\seealso{
\code{\link{regionplot}}
}
