% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionplot.R
\name{styleSignal}
\alias{styleSignal}
\title{S4 class containing the data to draw a regional association plot with the 
  'signal' style.}
\usage{
styleSignal(pValues, priorsd, priorc, cs_size)
}
\description{
S4 class containing the data to draw a regional association plot with the 
  'signal' style.
}
\section{Slots}{

\describe{
\item{\code{signalPValues}}{Data frame with p-values for the 'signal' style.}
}}

\seealso{
\code{\link{regionplot}}
}
