% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionplot.R
\name{styleSignals}
\alias{styleSignals}
\title{S4 class containing the data to draw a regional association plot with the 
  'signals' style.}
\usage{
styleSignals(pValues, analysis, chrom, pos_start, pos_end, priorsd, priorc,
  cs_size)
}
\description{
S4 class containing the data to draw a regional association plot with the 
  'signals' style.
}
\section{Slots}{

\describe{
\item{\code{signalsPValues}}{Data frame containing the p-values for the 'signals' 
style.}

\item{\code{index_cleo}}{UNDEFINED}
}}

\seealso{
\code{\link{regionplot}}
}
