\name{text2factor}
\alias{text2factor}
\title{Convert text to factor}
\description{
  Convenience function to convert test to factor, with empty strings
  converted to NA.
}
\usage{
text2factor(t)
}
\arguments{
  \item{t}{a character vector}
}
\value{
  A factor vector.
}
\author{
  Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
