% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{validate_exposure_qtl}
\alias{validate_exposure_qtl}
\title{Validate an input for instrument extraction of molecular traits}
\usage{
validate_exposure_qtl(analysis, entities)
}
\arguments{
\item{analysis}{(required): a keyword/keyphrase to search for the study of interest}

\item{entities}{(default): other parameters}
}
\description{
Validate an input for instrument extraction of molecular traits
}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
\keyword{internal}
