% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_connections.R
\name{validate_impala}
\alias{validate_impala}
\title{Validate and/or establish an impala odbc connection}
\usage{
validate_impala(impala = getOption("gtx.impala", NULL))
}
\arguments{
\item{impala}{impala connection to validate. Defaults to checking options(gtx.impala).}
}
\value{
implyr impala connection
}
\description{
\strong{validate_impala() - Validate impala connection}
This will validate an implyr impala connection. If it does not 
exists, it will try to safely establish the the connection 
and return it.
}
\examples{
impala <- validate_impala()
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
