% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{validate_module_input}
\alias{validate_module_input}
\title{Validate module input has been changed}
\usage{
validate_module_input(...)
}
\arguments{
\item{...}{A single or multiple comma seperated inputs to validate.}
}
\value{
Returns nothing. Will fail if encounters a problem.
}
\description{
\strong{validate_module_input - validate module input}
In the gwas_results modules, many inputs are set to "DOUBLE_CLICK_HERE_TO_CHANGE".
This functions validates that the input was changed from the default. 
It is important to wrap this functionality in a function to keep it hidden from users.
}
\examples{
validate_module_input(hgncid)
validate_module_input(chrom, pos_start, pos_end)
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
