% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mranalysis.R
\name{visual_for_mr}
\alias{visual_for_mr}
\title{Visualisation of MR results}
\usage{
visual_for_mr(mrres, scatter = TRUE, forest = TRUE, funnel = TRUE,
  exp = "internal", out = "internal",
  dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{mrres}{(required): the list of dataframes from run_for_mr method}

\item{scatter}{(optional): if TRUE creates a scatter plot for an exposure/outcome pair}

\item{forest}{(optional): if TRUE creates a forest plot for an exposure/outcome pair}

\item{funnel}{(optional): if TRUE creates a funnel plots for an exposure/outcome pair}

\item{exp}{(optional): if "internal", the plot labels will be formatted assuming the exposure is 
from the internal source; it should be changed to "external" wherever is applicable.}

\item{out}{(optional): if "internal", the plot labels will be formatted accordingly for the outcome
as for the exposure; it should be changed to "external" wherever is applicable.}
}
\value{
the list of three types of plots for one exposure vs one outcome:
1) scatter plot
2) forest plot
3) funnel plot
}
\description{
This methods creates three types of plots for a chosen exposure and outcome.
}
\examples{
To access the plots, e.g. \dontrun{plots[[1]]} - a scatter plot

}
\author{
Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
}
