% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_connections.R
\name{whoami}
\alias{whoami}
\title{whoami - Determine user name/id for working with CDH.}
\usage{
whoami()
}
\description{
whoami - Determine user name/id for working with CDH.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
