% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.plot}
\alias{aba.plot}
\title{\strong{aba.plot() - Plot aba results}}
\usage{
aba.plot(.data, p12_ge = 0.8, max_dot_size = 5, title = NULL,
  flog_level = getOption("gtx.flog_level", "WARN"))
}
\arguments{
\item{p12_ge}{[Default = 0.80] Values below this will be colored as non-significant, while above have color for direction.}

\item{max_dot_size}{[Default = 5] Adjust max size of dots.}

\item{title}{Set the plot title. NULL will remove title.}

\item{flog_level}{[Default = getOption("gtx.flog_level", "WARN")] \code{\link{futile.logger}} \code{\link{flog.threshold}} [INFO, WARN, ERROR]}
}
\value{
ggplot2 object for viz and export.
}
\description{
This function tries to create a bubble plot to 
illustrate the coloc region. This function returns
a ggplot object which can be viewed or saved. It 
is \emph{highly} recommended to save the plot to pdf 
for best visualization of the plot.
}
\examples{
\\strong{Basic use:}
coloc_plot <- \\code{\\link{aba.plot}}(colocs_filtered)
coloc_plot <- \\code{\\link{aba.plot}}(colocs_filtered, max_dot_size = 3, title = "Cool plot")

\\strong{View the returned object:}
\\code{coloc_plot}

\\strong{Save the object to PDF:}
\\code{ggsave(filename = "hgmcr_region_coloc.pdf", 
       plot     = coloc_plot, 
       width    = 11, 
       height   = 8.5, 
       dpi      = 300)}
       
\\strong{Advanced use: query data, remove death related traits, filter, and then plot}
\\code{coloc_plot <- aba.query_locus(hgncid = "HMGCR", sc = sc) \%>\% 
                filter(!str_detect(description, "death")) \%>\% 
                aba.filter() \%>\% 
                aba.plot(title = "Best plot ever")}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
