% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba.R
\name{aba.wrapper}
\alias{aba.wrapper}
\title{aba.wrapper - single function to query and plot the aba colocs}
\usage{
aba.wrapper(analysis_ids, hgncid, ensemblid, rsid, chrom, pos, pos_start,
  pos_end, impala = getOption("gtx.impala", NULL), ...)
}
\arguments{
\item{analysis_ids}{ukbiobank analysis id(s), single string or vector.}

\item{hgncid}{HGNC symbol. Single string or vector.}

\item{ensemblid}{Ensembl gene ID. Single string or vector.}

\item{rsid}{SNP rsid. Single string or vector.}

\item{chrom}{chromosome - Used to define a specific region}

\item{pos}{position that will have +/- \code{surround} for bounds to pull in colocs.}

\item{pos_start}{start position - Used to define a specific region, overrides surround}

\item{pos_end}{end position - Used to define a specific region, overrides surround}

\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}
}
\value{
data.frame with the inputs used, all the data for each input, and default plots
}
\description{
aba.wrapper - single function to query and plot the aba colocs
}
\examples{
colocs <- aba.wrapper(hgncid = "HMGCR")
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
