\name{clinical.import}
\alias{clinical.import}
\title{Import multiple files exported from SAS}
\description{
  Provides a convenience function to build a list of data frames,
  where each data frame is generated by reading a .txt file assumed
  to have been generated by SAS proc export.
}
\usage{
clinical.import(d, pattern = "^[a-zA-Z][a-zA-Z1-9]*\\\\.txt", 
                usubjid = getOption("gtx.usubjid", "USUBJID"), 
                verbose = TRUE, convert.YN = TRUE, convert.Date = TRUE, 
                only)
}
\arguments{
  \item{d}{Path to directory containing files exported from SAS}
  \item{pattern}{Regular expression for files to be imported}
  \item{usubjid}{Name of variable used for unique subject identifier}
  \item{verbose}{Whether to print progress messages}
  \item{convert.YN}{Whether to convert columns of Y and N to logical type}
  \item{convert.Date}{Whether to convert ddMMMyyyy columns to R Date type}
  \item{only}{Character vector}
}
\details{
  \code{clinical.import} provides a convenient method to read in
  clinical data from one or more .txt files generated by SAS proc
  export.

  The convenience features are that (i) data from multiple files are
  imported into a single object, a list of data frames, with one data
  frame for each imported file; (ii) data coded as Y/N and as text
  dates, are converted to appropriate R types; (iii) text data are
  imported as factors except for subject identifiers, which are imported
  as text; (iv) factor levels for text data are converted when necessary
  from latin-9 encoding (used by SAS proc export) to UTF-8 encoding.

  These features are intended to make the imported data structure work
  well with other gtx functions including \code{\link{clinical.derive}()}
  and \code{link{demographics}()}.

  By default, all files matching \code{pattern} in directory \code{d}
  are imported into the list of data frames.  The default for
  \code{pattern} is to exclude files with underscores because the
  authors work in an environment where the SAS exported data in .txt
  files is accompanied by metadata in corresponding _spec.txt files.
  This behaviour can be changed but the function will only work if all
  the files targeted have .txt extensions.  The \code{only} argument
  allows a subset of files to be read, the arguments to \code{only}
  should be the names of files without the .txt extensions.

  The variables named \code{usubjid} is imported as class character, all
  other character variables are imported as factors.
}
\value{
  A list of dataframes.
}
\examples{
\dontrun{
clindata <- clinical.import("path/to/clinical/export/")
data(derivations.standard)
gxvars <- clinical.derive(clindata, derivations.standard)
summary(gxvars)
}
}
\author{
  Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
