\name{derivations.standard}
\alias{derivations.standard}
\docType{data}
\title{Exploratory rules for deriving variables from clinical data}
\description{
  Some rules for deriving variables for genetic association analysis
  from clinical data, for use with clinical.derive and associated functions.
}
\usage{data(derivations.standard)}
\format{
  A data frame.

  You should not manipulate this data frame directly.  Instead, use \code{\link{derivation.add}()}.
}
\examples{
data(derivations.standard)
}
\keyword{datasets}
