% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionplot.R
\name{getDataForRegionplot}
\alias{getDataForRegionplot}
\title{Get data for regionplot.}
\usage{
getDataForRegionplot(analysis, entity, signal, chrom, pos_start, pos_end,
  pos, hgncid, ensemblid, rs, surround = 5e+05, style = "signals",
  priorsd = 1, priorc = 1e-05, cs_size = 0.95, maf_ge, rsq_ge,
  emac_ge, case_emac_ge, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{analysis}{Character giving the key value of the GWAS to draw a regional
association plot from.}

\item{entity}{Entity within eQTL to draw the regional association plot for, 
as ENSEMBL or HGNC identifier. Character.}

\item{signal}{}

\item{chrom}{Character specifying the chromosome.}

\item{pos_start}{The start position of the region for which to draw the plot.
Integer.}

\item{pos_end}{The end position for which to draw the plot. Integer.}

\item{pos}{The position for which to define the region around. Integer.}

\item{hgncid}{The HGNC identifier of the gene to define the region around.
Character.}

\item{ensemblid}{The ENSEMBL gene identifier to define the region around. 
Character.}

\item{rs}{The dbSNP RS identifier of the variant to define the region around.
Character.}

\item{surround}{The distance around the gene to include in the region for 
plotting. Default is \code{500000}.}

\item{style}{Character vector specifying the style or styles of plot to 
output. Can be one or more of \code{c('none', 'signals', 'signal', 
'classic', 'ld')}.}

\item{priorsd}{Default is \code{1}. Numeric.}

\item{priorc}{Default is \code{1e-5}. Numeric.}

\item{cs_size}{Default is \code{0.95}. Numeric.}

\item{maf_ge}{Numeric filtering threshold. Minor alleles with a frequency 
greater than or equal to \code{maf_ge} will be included in the plot.}

\item{rsq_ge}{Numeric filtering threshold, specifying imputation R-squared 
greater than or equal to \code{rsq_ge}.}

\item{emac_ge}{Numeric filtering threshold. Minor allele count must be 
greater than or equal to \code{emac_ge}.}

\item{case_emac_ge}{Numeric filtering threshold. Case minor allele count must
be greater than or equal to \code{case_emac_ge}.}

\item{dbc}{A database connection.}
}
\value{
A \code{\linkS4class{regionplot}} object containing the data needed 
  to draw a regional association plot.
}
\description{
\code{getDataForRegionplot()} returns all of the data required to draw a
  regional association plot with the \code{regionplot()} function.
}
\seealso{
\code{\link{regionplot()}} to get the data and draw regional 
  association plots in one step.
}
\concept{\code{\link{regionplot}} functions.}
