\name{gtx-package}
\alias{gtx-package}
\alias{gtx}
\docType{package}
\title{Genetics ToolboX}
\description{
  This package implements assorted tools for genetic association
  analyses, which is viewed as being entirely an exercise in regressing
  a (possibly multivariate) phenotypic \dQuote{response variable} onto
  one or more \dQuote{explanatory variables} that include genetic
  variables.

  Currently, this package does not provide computationally efficient
  functions for genetic association analyses at a genome wide scale
  (genome wide association studies; GWAS).  These are already provided
  by other R packages and by standalone software such as PLINK.  Rather,
  the focus of this package is to provide functions for analysing and
  manipulating phenotype data before conducting a GWAS
  (\dQuote{pre-GWAS}), and on functions for analysing summary statistics
  resulting from a GWAS (\dQuote{post-GWAS}).  Many of the
  \dQuote{post-GWAS} functions implement regression analyses using
  summary statistics, which are intended to closely approximate results
  that would be obtained by directly analysing the subject-specific
  genotype and phenotype data.

  Functions for \dQuote{pre-GWAS} analyses include functions useful for
  deriving response variables from phenotype data, especially response
  variables for pharmacogenetic analyses derived from clinical trial
  phenotype data; functions for power analyses; and functions for
  annotating and plotting results.
  
  Functions for \dQuote{post-GWAS} analyses currently support
  calculation of approximate Bayes factors; multi-SNP risk score
  analyses; multi-SNP conditional regression analyses; and
  multi-phenotype analyses.

  Approximate Bayes factors can be calculated using
  \code{\link{abf.Wakefield}}, \code{\link{abf.normal}} and
  \code{\link{abf.t}}.
  
  For multi-SNP risk score analyses, the main functions for analysing
  summary statistics are \code{\link{grs.summary}},
  \code{\link{grs.plot}} and \code{\link{grs.filter.Qrs}}.  The summary
  statistics necessary for these analyses are single SNP association
  statistics, which can be calculated using a wide variety of existing
  tools for GWAS analysis and meta-analysis.
  
  For multi-SNP conditional or multiple regression analyses, the main
  functions for performing multiple regression using summary statistics
  are \code{\link{combine.moments2}}, \code{\link{est.moments2}},
  \code{\link{lm.moments2}} and \code{\link{stepup.moments2}}.  The
  summary statistics necessary for these analyses can be calculated from
  subject-specific genotype and phenotype data, using the function
  \code{\link{make.moments2}}.

  Multi-phenotype analyses can be performed using
  \code{\link{multipheno.T2}} and \code{\link{multipheno.OBrien}}.
  
  In addition, there are \dQuote{helper} functions for reading and
  manipulating subject-specific genotype and phenotype data, and which
  provide a convenient interface from R to genotype data exported from
  PLINK, and imputed genotype data generated by MACH, minimac, or
  IMPUTE.  These provide a platform for calculating the necessary
  summary statistics, and for performing \dQuote{exact} analyses to
  validate some of the approximate summary statistic based methods.  The
  main functions provided are \code{\link{read.snpdata.plink}},
  \code{\link{read.snpdata.mach}}, \code{\link{read.snpdata.minimac}},
  and \code{\link{read.snpdata.impute}}.
}

\author{Toby Johnson
\email{Toby.x.Johnson@gsk.com}}
