% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtxbase.R
\name{gtxanalyses}
\alias{gtxanalyses}
\title{Find GWAS analyses}
\usage{
gtxanalyses(analysis, analysis_not, phenotype_contains, description_contains,
  has_tag, ncase_ge, ncohort_ge, has_cleo = FALSE,
  analysis_fields = c("description", "phenotype", "covariates", "cohort",
  "unit", "ncase", "ncontrol", "ncohort"), tag_is, with_tags = FALSE,
  has_access_only = FALSE, dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{analysis}{Key value(s) for GWAS analysis/es required}

\item{analysis_not}{Key value(s) for GWAS analysis/es not required}

\item{phenotype_contains}{String to search for in phenotype metadata}

\item{description_contains}{String to search for in description metadata}

\item{has_tag}{Tag key value(s) for GWAS analysis/es required}

\item{ncase_ge}{Numeric to select number of cases greater-or-equal}

\item{ncohort_ge}{Numeric to select number in cohort greater-or-equal}

\item{has_cleo}{Logical, select only analyses with \link{CLEO} results}

\item{analysis_fields}{Names of columns/fields to return}

\item{tag_is}{Internal use only}

\item{with_tags}{Internal use only}

\item{has_access_only}{Deprecated}

\item{dbc}{Database connection}
}
\value{
A data frame of selected rows and columns from \code{TABLE analyses}.
}
\description{
Find GWAS analyses matching specified criteria.
}
\details{
In the near future, a better high level interface will be 
provided to find GWAS analyses of interest.  The \code{gtxanalyses()} 
and \code{gtxwhat()} functions were originally intended as internal
functions, to abstract part of the functionality required within the
\code{phewas()} family of functions.  However, in the absence of a
better interface, these functions are currently exported for users to
use.
}
\author{
Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
