% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{gwas}
\alias{gwas}
\title{Functions for reporting and visualizing a complete gwas}
\usage{
gwas(analysis, style = c("manhattan", "qqplot"), pval_thresh = 5e-08,
  maf_ge, rsq_ge, emac_ge, case_emac_ge, gene_annotate = TRUE,
  plot_ymax = 30, prune_dist = 500000L, manhattan_thresh = 5e-08,
  manhattan_col = c("#064F7C", "#6D97BD"),
  manhattan_interspace = 5e+07, qqplot_col = "#064F7C",
  qqplot_alpha = 0.01, plot_fastbox = 2, zrok = FALSE,
  dbc = getOption("gtx.dbConnection", NULL))
}
\arguments{
\item{gene_annotate}{(logical): whether to annotate top hits with gene names}

\item{analysis:}{The key value for the GWAS analysis to summarize}

\item{style:}{Character vector specifying plot style(s). Can be 'manhattan'
or 'qqplot'. style = 'none' will produce no plot.}

\item{pval_thresh:}{P-value threshold for top hits table}

\item{maf_ge:}{MAF greater-or-equal threshold}

\item{rsq_ge:}{Imputation r-squared greater-or-equal threshold}

\item{emac_ge:}{UNDEFINED}

\item{case_emac_ge:}{UNDEFINED}

\item{plot_ymax:}{Y axis maximum, -log10(p) scale, for plots}

\item{prune_dist:}{UNDEFINED}

\item{manhattan_thresh:}{P-value for threshold line on Manhattan plot}

\item{manhattan_col:}{Chromosome colour cycle for Manhattan plot}

\item{manhattan_interspace:}{Chromosome interspacing for Manhattan plot}

\item{qqplot_col:}{Colour for QQ plot}

\item{qqplot_alpha:}{Null envelope alpha value for QQ plot}

\item{plot_fastbox:}{Box height for fast approximate plotting}

\item{zrok:}{UNDEFINED}

\item{dbc:}{Database connection}

\item{connectionType}{(character): The type of connection. Only SQL allowed for now}
}
\value{
Data frame of pruned data and pplot of style 'manhattan' and/or 'qqplot'
}
\description{
This high level function summarizes and visualizes results
(i.e. complete summary statistics) from a single GWAS analysis.
Currently, a distance-pruned table of top hits, and optional
Manhattan and QQ plots, are generated.
The distance pruning and gene annotation is currently very slow and
needs to be optimized.
The Manhattan and QQ plots use an approximation to speed up both the data
query and the plot rendering, where boxes or lines are drawn instead of
individual points, for y-axis values between 0 and
\code{plot_fastbox}.  Faint grey hatching is shown over this part of
the plot to indicate that the true data are not being plotted.
Y-axis points are truncated, but this needs to be made more clear in the
plot.
}
