% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_gwas_summarize}
\alias{ideas_gwas_summarize}
\title{ideas_gwas_summarize}
\usage{
ideas_gwas_summarize(.data, relaxed = 3, group = FALSE, n_cs_ge = 10,
  n_snps_ge = 10, impala = getOption("gtx.impala", NULL))
}
\arguments{
\item{.data}{\code{\link{ideas_predict}} output}

\item{relaxed}{[3] Numeric value, 1:10, rank cutoff_gt for inclusion. e.g. 3 = Must be in top 3 cell types.}

\item{group}{[FALSE] TRUE = Group similar cell types.}

\item{n_cs_ge}{[10] min # (>=) of cred sets to GWAS summarize}

\item{n_snps_ge}{[10] min # (>=) of snps across all cred sets}
}
\description{
Summarize most likely cell type per GWAS analysis id.
}
\seealso{
Other ideas_predict: \code{\link{ideas_gwas_plot}},
  \code{\link{ideas_loci_summarize}},
  \code{\link{ideas_make}}, \code{\link{ideas_predict}},
  \code{\link{ideas_preload_states}},
  \code{\link{ideas_wrapper}}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
\concept{ideas_predict}
