% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_lasso}
\alias{ideas_lasso}
\title{ideas_lasso}
\usage{
ideas_lasso(chrom, pos, vindex, vprior, vlp, states_data = NULL,
  use_genome_background = TRUE, inter = 5, alpha = 0.99, layer = 10,
  permute = FALSE, states = NULL, impala = getOption("gtx.impala", NULL),
  db = config_db())
}
\arguments{
\item{chrom}{Chromosomes}

\item{pos}{Positions}

\item{vindex}{TBD}

\item{vprior}{TBD}

\item{vlp}{TBD}

\item{use_genome_background}{[TRUE]}

\item{inter}{[5]}

\item{alpha}{[0.99]}

\item{layer}{[10]}

\item{permute}{[FALSE]}

\item{states}{TBD}

\item{impala}{[getOption("gtx.impala", NULL)] Implyr impala connection}
}
\value{
TBD
}
\description{
Using gglasso, identify the most likely cell type per locus.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
