% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_make}
\alias{ideas_make}
\title{Ideas make}
\usage{
ideas_make(analysis, pval_le = 5e-08, chrom, pos, case_emac_ge = 25)
}
\arguments{
\item{analysis}{GWAS analysis id.}

\item{pval_le}{[5e-8] Max pval for GWAS}

\item{chrom}{Chromosome to query}

\item{pos}{Position}

\item{case_emac_ge}{[25]}
}
\value{
Table of GWAS loci with posteriors for \code{\link{ideas_predict}}
}
\description{
Convert GWAS summary stats into posterior probabilities that can be used by IDEAS predict.
}
\seealso{
Other ideas_predict: \code{\link{ideas_gwas_plot}},
  \code{\link{ideas_gwas_summarize}},
  \code{\link{ideas_loci_summarize}},
  \code{\link{ideas_predict}},
  \code{\link{ideas_preload_states}},
  \code{\link{ideas_wrapper}}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
\concept{ideas_predict}
