% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_predict}
\alias{ideas_predict}
\title{ideas_predict}
\usage{
ideas_predict(.data, states = NULL, permute = FALSE, inter = 5,
  binsz = 200, pcut = 6, CI = 0.95, seed = NULL)
}
\arguments{
\item{permute}{[Default = FALSE]}

\item{inter}{[Default = 5]}

\item{binsz}{[Default = 200]}

\item{pcut}{[Default = 6]}

\item{CI}{[Default = 0.95]}

\item{seed}{[Default = NULL]}

\item{alpha}{[Default = 0.99]}

\item{layer}{[Default = 10]}
}
\value{
TBD
}
\description{
Predict IDEAS states from input GWAS data supplied by \code{ideas_make}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
