% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_predict.R
\name{ideas_predict}
\alias{ideas_predict}
\title{ideas_predict}
\usage{
ideas_predict(.data, states_data = NULL, states = NULL,
  permute = FALSE, inter = 5, binsz = 200, pcut = 6, CI = 0.95,
  seed = NULL, db = config_db())
}
\arguments{
\item{.data}{\code{\link{ideas_make}} input data}

\item{states}{[NULL] TODO}

\item{permute}{[FALSE] TODO}

\item{inter}{[5] TODO}

\item{binsz}{[200] Padding around each cred set SNP.}

\item{pcut}{[6] TODO}

\item{CI}{[0.95] Cred set confidence interval.}

\item{seed}{[NULL] Set seed for RNG number generator.}

\item{db}{= [`gtx::config_db()`] Database contained IDEAS states data.}

\item{alpha}{[0.99] TODO}

\item{layer}{[10] TODO}

\item{permute}{[FALSE] TODO}

\item{states}{Select subset of states. Not currently used as of 2019-01-11.}
}
\description{
Predict IDEAS states from input GWAS data supplied by \code{\link{ideas_make}}
}
\seealso{
Other ideas_predict: \code{\link{ideas_gwas_plot}},
  \code{\link{ideas_gwas_summarize}},
  \code{\link{ideas_loci_summarize}},
  \code{\link{ideas_make}},
  \code{\link{ideas_preload_states}},
  \code{\link{ideas_wrapper}}
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}

Karl Guo \email{karl.x.guo@gsk.com}
}
\concept{ideas_predict}
