% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_connections.R
\name{impala_copy_to}
\alias{impala_copy_to}
\title{Copy data to user's home database for joins with impala tables.}
\usage{
impala_copy_to(df, dest = getOption("gtx.impala", NULL),
  database = gtx::config_tmp_write_db(),
  table_name = glue("tmp_{gtx::whoami()}_data4join"), random_name = TRUE)
}
\arguments{
\item{df}{Data to copy to RDIP}

\item{dest}{Impala implyr connection}

\item{database}{Defaults to user's database: \code{whoami}.}

\item{table_name}{[Default = "tmp_data4join"] Specify table name, overrides random_name = TRUE}

\item{random_name}{[Default = FALSE] TRUE = ~random number name for the table.}
}
\value{
table reference to the \code{df} in RDIP
}
\description{
\strong{impala_copy_to() - copy data to RDIP}
This will copy a dataframe to the home directory of the user 
and return the
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
