% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_connections.R
\name{impala_copy_to}
\alias{impala_copy_to}
\title{Copy data to user's home database for joins with impala tables.}
\usage{
impala_copy_to(df, dest = getOption("gtx.impala", NULL))
}
\arguments{
\item{df}{Data to copy to RDIP}

\item{dest}{Impala implyr connection}
}
\value{
table reference to the \code{.data} in RDIP
}
\description{
\strong{impala_copy_to() - copy data to RDIP}
This will copy a dataframe to the home directory of the user 
and return the
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
