% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{int_ht_cred_set_wrapper}
\alias{int_ht_cred_set_wrapper}
\title{int_ht_cred_set_wrapper}
\usage{
int_ht_cred_set_wrapper(analysis, chrom, pos, style, ...)
}
\arguments{
\item{analysis}{analysis ID}

\item{chrom}{chromosome}

\item{pos}{position}

\item{...}{Additional arguements for regionplot.data}
}
\value{
tibble
}
\description{
int_ht_cred_set_wrapper
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
