% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{int_ht_regional_context}
\alias{int_ht_regional_context}
\title{int_ht_regional_context}
\usage{
int_ht_regional_context(input, chrom, pos, analysis, signal, cpu = 8,
  drop_cs = TRUE, dbc = getOption("gtx.dbConnection", NULL), ...)
}
\arguments{
\item{input}{data.frame with chrom, pos, analysis, signal; likely & suggested input from int_ht_phewas. Other cols will be kept.}

\item{chrom}{chromosome (Mandatory)}

\item{pos}{position (Mandatory)}

\item{analysis}{analysis ID (Mandatory)}

\item{signal}{CLEO signal OR NA. (Mandatory)}

\item{dbc}{[Default = getOption("gtx.dbConnection", NULL)] gtxconnection.}

\item{...}{Params to pass to regionplot.data (e.g. surround = 1e6)}
}
\value{
tibble
}
\description{
High-Throughput regional_context.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
