% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_context.R
\name{regional_context_analysis}
\alias{regional_context_analysis}
\title{regional_context_analysis}
\usage{
regional_context_analysis(hgnc, hgncid, rs, rsid, chrom, pos, ref, alt,
  ignore_ukb_neale = TRUE, ignore_ukb_cane = TRUE,
  ignore_qtls = TRUE, phewas_pval_le = 1e-07, cpu = 8,
  drop_cs = TRUE, dbc = getOption("gtx.dbConnection", NULL), ...)
}
\arguments{
\item{hgnc}{Analyze all moderate/high impact VEP variants overlapping gene position.}

\item{rs}{RSID(s) to analyze as a string or vector}

\item{chrom}{chromosome(s) to analyze as a string or vector}

\item{pos}{positions(s) to analyze as a single or vector}

\item{ref}{ref allele(s) to analyze as a string or vector. Optional w.r.t. chrom/pos.}

\item{alt}{alt allele(s) to analyze as a string or vector. Optional w.r.t. chrom/pos.}

\item{ignore_ukb_neale}{[Default = TRUE] TRUE = ignore Neale UKB GWAS in PheWAS}

\item{ignore_ukb_cane}{[Default = TRUE] TRUE = ignore Canelaxandri UKB GWAS in PheWAS}

\item{ignore_qtls}{[Default = TRUE] TRUE = ignore QTLs in PheWAS (gwas_results with entity)}

\item{cpu}{[Default = 8] Number of cpus to use for regional context analysis (i.e. regionplot.data's)}

\item{dbc}{[Default = getOption("gtx.dbConnection", NULL)] gtxconnection.}

\item{...}{regionplot.data params}
}
\value{
tibble (data.frame)
}
\description{
Calculate regional_context on-the-fly. 
*Critical - DO NOT run this function in a loop, use vector inputs*.
Run PheWAS for all variants, identifying GWAS at the significance threshold
(1e-7 by default). If a gene is used as input, the function will find all moderate/high impact
variants from the VEP table and use all the VEP variants for PheWAS. 
In the Phewas, the default behavior is to also ignore Ben Neale v1 & Canelaxandri17 UKB,
and all xQTL data. After the PheWAS, each positive variant+GWAS pair is clumped by GWAS top hits (th_pos). 
Next, for each variant+GWAS pair, generate credible sets and determine if the variant is in the credible set and gather 
others stats on each credible set. Perform this analysis on both marginal GWAS and CLEO GWAS results.
This function will also use multiple cores if available.
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
