% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionplot.R
\name{regionplot}
\alias{regionplot}
\title{S4 class containing the data to draw a regional association plot.}
\usage{
regionplot(analysis, entity, signal, chrom, pos_start, pos_end, pos,
  hgncid, ensemblid, rs, surround = 5e+05, maf_ge, rsq_ge, emac_ge,
  case_emac_ge, priorsd = 1, priorc = 1e-05, cs_size = 0.95,
  plot_ymax = 300, style = "signals", protein_coding_only = TRUE,
  highlight_style = "circle", dbc = getOption("gtx.dbConnection",
  NULL))

regionplot(analysis, entity, signal, chrom, pos_start, pos_end, pos,
  hgncid, ensemblid, rs, surround = 5e+05, maf_ge, rsq_ge, emac_ge,
  case_emac_ge, priorsd = 1, priorc = 1e-05, cs_size = 0.95,
  plot_ymax = 300, style = "signals", protein_coding_only = TRUE,
  highlight_style = "circle", dbc = getOption("gtx.dbConnection",
  NULL))
}
\arguments{
\item{analysis}{Character giving the key value for the GWAS analysis to draw 
a regional association plot from.}

\item{entity}{Optional entity within eQTL to draw the regional association 
plot for, as ENSEMBL or HGNC identifier. Character.}

\item{signal}{}

\item{chrom}{Optional character specifying the chromosome.}

\item{pos_start}{Optional start position of the region for which to draw the 
plot. Integer.}

\item{pos_end}{Optional end position of the region for which to draw the 
plot. Integer.}

\item{pos}{Optional position for which to define the region around. Integer.}

\item{hgncid}{Optional HGNC identifier of the gene to define the region 
around. Character.}

\item{ensemblid}{Optional ENSEMBL gene identifier to define the region 
around. Character.}

\item{rs}{dbSNP RS identifier of the variant to define the region around.
Character.}

\item{surround}{Distance around the gene to include in the region for 
plotting. Default is \code{500000}.}

\item{maf_ge}{Numeric filtering threshold. Minor alleles with a frequency 
greater than or equal to \code{maf_ge} will be included in the plot.}

\item{rsq_ge}{Numeric filtering threshold, specifying imputation R-squared 
greater than or equal to \code{rsq_ge}.}

\item{emac_ge}{Numeric filtering threshold. Minor allele count must be 
greater than or equal to \code{emac_ge}.}

\item{case_emac_ge}{Numeric filtering threshold. Case minor allele count must
be greater than or equal to \code{case_emac_ge}.}

\item{priorsd}{Optional. Default is \code{1}. Numeric.}

\item{priorc}{Optional. Default is \code{1e-5}. Numeric.}

\item{cs_size}{Default is \code{0.95}. Numeric.}

\item{plot_ymax}{Numeric maximum y-axis value. Default is \code{300}.}

\item{style}{Character vector specifying the style or styles of plot to 
output. Can be one or more of \code{c('none', 'signals', 'signal', 
'classic', 'ld')}. Default is \code{'signals'}.}

\item{protein_coding_only}{Whether to restrict annotation to protein coding
genes only. Default is \code{TRUE}.}

\item{highlight_style}{Highlighting style for plotting. Default is 
\code{'circle'}.}

\item{dbc}{A database connection. Default is 
\code{getOption("gtx.dbConnection", NULL)}.}
}
\value{
Regional association plots and a \code{\linkS4class{regionplot}} 
  object containing the data used to draw the regional association plot.
}
\description{
Draw a regional association plot.
}
\details{
\code{regionplot()} draws a regional association plot by querying the 
  required data via a database connection (see \code{\link{gtxdbcheck}}). 
  
  The region to plot results over can be specified in several different ways.
  The region can be supplied as physical coordinates using the arguments 
  \code{chrom}, \code{pos}, \code{pos_start} and \code{pos_end}. 
  Alternatively, the region can be centered on a gene of interest, using 
  either the \code{hgncid} or \code{ensemblid} argument. The size of the 
  region around the gene can be modified using the \code{surround} argument.
}
\section{Slots}{

\describe{
\item{\code{queryForPValues}}{The query used to get the p-values for a given GWAS 
from a database.}

\item{\code{pValues}}{A database containing the p-values for plotting.}

\item{\code{xEntity}}{A named list containing \code{entity}, \code{entity_label} and 
\code{entity_where}.}

\item{\code{xRegion}}{A named list containing \code{chrom}, \code{pos_start}, 
\code{pos_end} and \code{label}. 
\itemize{
 \item{\code{chrom} is the chromosome to be plotted.} 
 \item{\code{pos_start} and \code{pos_end} specify the start and end of 
  the region to draw the plot for.} 
 \item{\code{label} gives a label for printing that shows the chromosome 
  and region.}
 }}

\item{\code{indexSNPs}}{A data frame with two columns for the position and p-values 
of SNPs that should be highlighted in the output plot.}

\item{\code{plotTitle}}{The title for a regional association plot.}

\item{\code{plotSubTitle}}{The subtitle for a regional association plot.}
}}

\seealso{
\code{\link{regionplot}}

\code{\link{getDataForRegionplot()}} to get the data for 
  \code{regionplot()}.
}
\author{
Toby Johnson \email{Toby.x.Johnson@gsk.com}
}
\concept{\code{\link{regionplot.data}, \link{getDataForRegionplot}}}
