% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{validate_module_input}
\alias{validate_module_input}
\title{Validate module input has been changed}
\usage{
validate_module_input(.data = NULL)
}
\arguments{
\item{.data}{A single/vector of inputs to validate.}
}
\value{
component \code{error} as TRUE or FALSE.
}
\description{
\strong{validate_module_input - validate module input}
In the gwas_results modules, many inputs are set to "DOUBLE_CLICK_HERE_TO_CHANGE".
This functions validates that the input was changed from the default. 
It is important to wrap this functionality in a function to keep it hidden from users.
}
\examples{
inputs <- validate_module_input(hgncid)
inputs <- validate_module_input(c(chrom, pos))
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
