% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_connections.R
\name{validate_sc}
\alias{validate_sc}
\title{Validate and/or establish a spark connection}
\usage{
validate_sc(sc = getOption("gtx.sc", NULL), spark_version = "2.2.0",
  app_name = "sparklyr", sc_config = spark_config())
}
\arguments{
\item{sc}{Spark connection to validate. Defaults to checking options(gtx.sc).}
}
\value{
Spark connection
}
\description{
\strong{validate_sc() - Validate spark connection}
This will validate a spark connection. If it does not 
exists, it will try to safely establish the the connection 
and return it. Note, if \code{\link{validate_sc}} has to 
create the Spark connection it will use the default settings.
}
\examples{
sc <- validate_sc(sc)
sc <- validate_sc()
}
\author{
Karsten Sieber \email{karsten.b.sieber@gsk.com}
}
